==> Data Analysis/4-2.AppPoor2.py <==
# -*- coding:utf-8 -*-

##############################################################################
# Part7：应用层次贫困生帮扶
##############################################################################

import pandas as pd
import json

# 读取原始的数据集
filepath_StudentsConsumption = '../education_data/7_consumption.csv'

filepath_StudentsInfo = '../education_data/2_student_info.csv'

filepath_StudentsConsumption_Processed = '../education_data/more/Student/5.Consumption.csv'


##############################################################################
# 完成贫困群体的消费规律，强调对比性，即疑似贫困生和正常消费水平的学生的消费习惯
# graph：1 学生消费总金额对比：统计一个月每一天的学生消费情况，选取贫困生与正常学生进行对比
# （强调贫困生的消费总金额很少）
# 采用散点图的形式展示，x轴为学生的ID，y轴为日期，散点的大小表示消费总额，超过阈值即为异常，采用不同颜色表达
# graph：2 学生每天消费次数对比：统计一个月每一天的学生消费次数，选取贫困生与正常学生进行对比
# （强调贫困生的消费次数稳定）
# 采用折线图的形式，展示20个学生的一个月的的消费次数，x轴为日期，y轴为消费次数，一条折线代表一个学生
# graph：3 学生平均每次消费金额对比
# （强调贫困生的每次消费金额少）
# 图2 和 图3 建议在同一个图中表达，x轴表示日期，y轴表示日均消费金额，一个折线代表一个学生的数据，底下为柱状图代表每一天的消费次数
# 同时折线图设置阈值，柱状图设置阈值，过高为红色，可以选取10个贫困生和10个正常消费的学生，比较

# 产生贫困生群体的消费习惯
# 选取20个贫困生代表群体，选取的原则是消费次数在300次左右，但是平均消费金额较低的学生

students_consumption = [[13012, 419.1, 56.1, 7.47, 0], [13564, 3272, 381, 8.59, 1], [13599, 1854.6000000000001, 199.10000000000002, 9.31, 0], [13685, 2039.4, 188.10000000000002, 10.84, 0], [13947, 3009, 377, 7.98, 1], [13948, 1971, 307, 6.42, 1], [13949, 2262, 273, 8.29, 1], [13950, 3330, 375, 8.88, 1], [13951, 2180.2000000000003, 255.20000000000002, 8.54, 0], [13952, 2682, 226, 11.87, 1], [13953, 2326.5, 292.6, 7.95, 0], [13955, 1640.1000000000001, 199.10000000000002, 8.24, 0], [13956, 12.100000000000001, 1.1, 11.0, 0], [13957, 2271.5, 211.20000000000002, 10.76, 0], [13958, 1475.1000000000001, 191.4, 7.71, 0], [13959, 2787.4, 317.90000000000003, 8.77, 0], [13960, 785.4000000000001, 72.60000000000001, 10.82, 0], [13961, 2359.5, 256.3, 9.21, 0], [13963, 2967, 418, 7.1, 1], [13964, 3173.5000000000005, 268.40000000000003, 11.82, 0], [13965, 2969, 341, 8.71, 1], [13966, 2193.4, 257.40000000000003, 8.52, 0], [13967, 9.9, 1.1, 9.0, 0], [13968, 3169.1000000000004, 370.70000000000005, 8.55, 0], [13969, 4283.400000000001, 399.3, 10.73, 0], [13970, 3840.1000000000004, 331.1, 11.6, 0], [13971, 519.2, 52.800000000000004, 9.83, 0], [13972, 2752.2000000000003, 305.8, 9.0, 0], [13973, 2455.2000000000003, 251.90000000000003, 9.75, 0], [13974, 1784.2, 180.4, 9.89, 0], [13975, 2124.1000000000004, 261.8, 8.11, 0], [13976, 1536.7, 163.9, 9.38, 0], [13977, 1812.8000000000002, 178.20000000000002, 10.17, 0], [13978, 2057.0, 180.4, 11.4, 0], [13979, 1988.8000000000002, 145.20000000000002, 13.7, 0], [13980, 2066.9, 205.70000000000002, 10.05, 0], [13981, 3064.6000000000004, 286.0, 10.72, 0], [13982, 2441, 291, 8.39, 1], [13983, 2970.0000000000005, 393.8, 7.54, 0], [13984, 2030.6000000000001, 250.8, 8.1, 0], [13985, 3932.5000000000005, 361.90000000000003, 10.87, 0], [13986, 2669.7000000000003, 242.00000000000003, 11.03, 0], [13987, 2818, 376, 7.49, 1], [13988, 2744.5, 299.20000000000005, 9.17, 0], [13989, 1881.0000000000002, 205.70000000000002, 9.14, 0], [13990, 2979, 404, 7.37, 1], [13991, 1049.4, 123.20000000000002, 8.52, 0], [13992, 1827, 307, 5.95, 1], [13993, 2785, 369, 7.55, 1], [13994, 1945.9, 192.50000000000003, 10.11, 0], [13995, 1828.2, 214.50000000000003, 8.52, 0], [13996, 1657.7, 155.10000000000002, 10.69, 0], [13997, 2297.9, 215.60000000000002, 10.66, 0], [13998, 1853.5000000000002, 200.20000000000002, 9.26, 0], [13999, 2313, 296, 7.81, 1], [14000, 2547.6000000000004, 234.3, 10.87, 0], [14001, 3714, 412, 9.01, 1], [14002, 2499.2000000000003, 250.8, 9.96, 0], [14003, 2539, 341, 7.45, 1], [14004, 2431.0, 287.1, 8.47, 0], [14005, 2181.3, 275.0, 7.93, 0], [14006, 1007.6000000000001, 119.9, 8.4, 0], [14007, 1333.2, 162.8, 8.19, 0], [14008, 2620, 322, 8.14, 1], [14009, 2196.7000000000003, 236.50000000000003, 9.29, 0], [14010, 4542, 343, 13.24, 1], [14011, 2601, 275, 9.46, 1], [14012, 1830.4, 169.4, 10.81, 0], [14013, 1426.7, 148.5, 9.61, 0], [14014, 2472.8, 314.6, 7.86, 0], [14015, 3150.4, 369.6, 8.52, 0], [14016, 2373.8, 269.5, 8.81, 0], [14017, 1886.5000000000002, 216.70000000000002, 8.71, 0], [14018, 3294, 424, 7.77, 1], [14019, 2204.4, 271.70000000000005, 8.11, 0], [14020, 2289.1000000000004, 224.4, 10.2, 0], [14021, 1867.8000000000002, 202.4, 9.23, 0], [14022, 1038.4, 134.20000000000002, 7.74, 0], [14023, 2811.6000000000004, 233.20000000000002, 12.06, 0], [14024, 3831.3, 371.8, 10.3, 0], [14025, 3227.4, 322.3, 10.01, 0], [14026, 3819, 396, 9.64, 1], [14027, 1987.7000000000003, 246.40000000000003, 8.07, 0], [14028, 2219.8, 254.10000000000002, 8.74, 0], [14029, 2179.1000000000004, 260.70000000000005, 8.36, 0], [14030, 3466, 382, 9.07, 1], [14032, 2280.3, 287.1, 7.94, 0], [14033, 2180.2000000000003, 275.0, 7.93, 0], [14034, 2244.0, 294.8, 7.61, 0], [14035, 2200.0, 224.4, 9.8, 0], [14036, 2425.5, 266.20000000000005, 9.11, 0], [14037, 1979, 339, 5.84, 1], [14038, 3745.5000000000005, 335.5, 11.16, 0], [14039, 1939.3000000000002, 196.9, 9.85, 0], [14040, 2695.0, 342.1, 7.88, 0], [14041, 2674.1000000000004, 240.9, 11.1, 0], [14042, 1763.3000000000002, 188.10000000000002, 9.37, 0], [14043, 1716.0000000000002, 210.10000000000002, 8.17, 0], [14044, 2540, 291, 8.73, 1], [14045, 2867, 341, 8.41, 1], [14047, 2677, 304, 8.81, 1], [14048, 2203.3, 188.10000000000002, 11.71, 0], [14049, 3157, 348, 9.07, 1], [14050, 337, 77, 4.38, 1], [14051, 2158.2000000000003, 227.70000000000002, 9.48, 0], [14052, 1409.1000000000001, 174.9, 8.06, 0], [14053, 2072.4, 233.20000000000002, 8.89, 0], [14054, 2737.9, 260.70000000000005, 10.5, 0], [14055, 1369, 197, 6.95, 1], [14056, 3505.7000000000003, 359.70000000000005, 9.75, 0], [14057, 3304.4, 346.5, 9.54, 0], [14058, 1254.0, 139.70000000000002, 8.98, 0], [14059, 2248.4, 248.60000000000002, 9.04, 0], [14060, 2900.7000000000003, 257.40000000000003, 11.27, 0], [14061, 1256.2, 150.70000000000002, 8.34, 0], [14062, 2337.5, 254.10000000000002, 9.2, 0], [14063, 1886.5000000000002, 173.8, 10.85, 0], [14064, 1625.8000000000002, 189.20000000000002, 8.59, 0], [14065, 2042.7000000000003, 187.00000000000003, 10.92, 0], [14066, 2339.7000000000003, 308.0, 7.6, 0], [14067, 3474.9, 419.1, 8.29, 0], [14068, 1510.3000000000002, 187.00000000000003, 8.08, 0], [14069, 2002.0000000000002, 228.8, 8.75, 0], [14070, 1758.9, 206.8, 8.51, 0], [14071, 2085.6000000000004, 224.4, 9.29, 0], [14072, 2550.9, 277.20000000000005, 9.2, 0], [14073, 2710.4, 280.5, 9.66, 0], [14074, 3732.3, 361.90000000000003, 10.31, 0], [14075, 3664.1000000000004, 336.6, 10.89, 0], [14076, 215.60000000000002, 13.200000000000001, 16.33, 0], [14077, 1701.7, 181.50000000000003, 9.38, 0], [14078, 1026.3000000000002, 110.00000000000001, 9.33, 0], [14079, 1543.3000000000002, 155.10000000000002, 9.95, 0], [14080, 1399.2, 188.10000000000002, 7.44, 0], [14081, 2435.4, 289.3, 8.42, 0], [14082, 2580.6000000000004, 228.8, 11.28, 0], [14083, 1780.9, 171.60000000000002, 10.38, 0], [14084, 2904.0000000000005, 277.20000000000005, 10.48, 0], [14085, 2366.1000000000004, 294.8, 8.03, 0], [14086, 3059.1000000000004, 331.1, 9.24, 0], [14087, 1606.0000000000002, 172.70000000000002, 9.3, 0], [14088, 2459.6000000000004, 231.00000000000003, 10.65, 0], [14089, 126.50000000000001, 16.5, 7.67, 0], [14090, 2843.5000000000005, 282.70000000000005, 10.06, 0], [14091, 1628.0000000000002, 161.70000000000002, 10.07, 0], [14092, 2454.1000000000004, 211.20000000000002, 11.62, 0], [14093, 1527.9, 178.20000000000002, 8.57, 0], [14094, 1182.5, 110.00000000000001, 10.75, 0], [14095, 1952.5000000000002, 157.3, 12.41, 0], [14096, 1978.9, 225.50000000000003, 8.78, 0], [14097, 3530, 409, 8.63, 1], [14098, 2250.6000000000004, 256.3, 8.78, 0], [14099, 4516, 486, 9.29, 1], [14100, 2656.5, 292.6, 9.08, 0], [14101, 1381.6000000000001, 179.3, 7.71, 0], [14102, 1959.1000000000001, 235.4, 8.32, 0], [14103, 3773.0000000000005, 403.70000000000005, 9.35, 0], [14104, 1950.3000000000002, 246.40000000000003, 7.92, 0], [14105, 1840.3000000000002, 199.10000000000002, 9.24, 0], [14106, 3125.1000000000004, 284.90000000000003, 10.97, 0], [14107, 2192.3, 256.3, 8.55, 0], [14108, 3071.2000000000003, 325.6, 9.43, 0], [14109, 3555.2000000000003, 352.0, 10.1, 0], [14110, 2191.2000000000003, 299.20000000000005, 7.32, 0], [14111, 1304.6000000000001, 129.8, 10.05, 0], [14112, 1689.6000000000001, 203.50000000000003, 8.3, 0], [14113, 1687.4, 173.8, 9.71, 0], [14114, 2560.8, 289.3, 8.85, 0], [14115, 2075.7000000000003, 246.40000000000003, 8.42, 0], [14116, 1567.5000000000002, 184.8, 8.48, 0], [14117, 1520.2, 163.9, 9.28, 0], [14118, 53.900000000000006, 7.700000000000001, 7.0, 0], [14119, 1776.5000000000002, 139.70000000000002, 12.72, 0], [14120, 1551.0000000000002, 132.0, 11.75, 0], [14121, 2730.2000000000003, 295.90000000000003, 9.23, 0], [14122, 2180, 290, 7.52, 1], [14123, 6267.8, 410.3, 15.28, 0], [14124, 4325.200000000001, 411.40000000000003, 10.51, 0], [14125, 2495.9, 247.50000000000003, 10.08, 0], [14126, 2506.9, 299.20000000000005, 8.38, 0], [14127, 1382.7, 115.50000000000001, 11.97, 0], [14128, 3238.4, 326.70000000000005, 9.91, 0], [14129, 1586.2, 149.60000000000002, 10.6, 0], [14130, 2381, 306, 7.78, 1], [14131, 1083.5, 110.00000000000001, 9.85, 0], [14132, 2773.1000000000004, 297.0, 9.34, 0], [14133, 2244.0, 245.3, 9.15, 0], [14134, 2301.2000000000003, 250.8, 9.18, 0], [14135, 2085, 298, 7.0, 1], [14136, 1976.7000000000003, 217.8, 9.08, 0], [14137, 2568.5, 273.90000000000003, 9.38, 0], [14138, 1082.4, 92.4, 11.71, 0], [14139, 4009, 476, 8.42, 1], [14140, 3306, 397, 8.33, 1], [14141, 3478, 388, 8.96, 1], [14142, 1764.4, 200.20000000000002, 8.81, 0], [14143, 2361.7000000000003, 295.90000000000003, 7.98, 0], [14144, 2294.6000000000004, 264.0, 8.69, 0], [14145, 1823.8000000000002, 226.60000000000002, 8.05, 0], [14146, 1401.4, 151.8, 9.23, 0], [14147, 2230.8, 260.70000000000005, 8.56, 0], [14149, 1439.9, 135.3, 10.64, 0], [14150, 1199.0, 93.50000000000001, 12.82, 0], [14151, 2459, 288, 8.54, 1], [14152, 3284, 399, 8.23, 1], [14153, 2794.0, 279.40000000000003, 10.0, 0], [14154, 1802.9, 189.20000000000002, 9.53, 0], [14155, 1234.2, 166.10000000000002, 7.43, 0], [14156, 2206.6000000000004, 236.50000000000003, 9.33, 0], [14157, 3185.6000000000004, 343.20000000000005, 9.28, 0], [14158, 2901, 333, 8.71, 1], [14159, 3184.5000000000005, 368.50000000000006, 8.64, 0], [14160, 2737.9, 338.8, 8.08, 0], [14161, 1936.0000000000002, 191.4, 10.11, 0], [14162, 2828.1000000000004, 267.3, 10.58, 0], [14163, 2242.9, 195.8, 11.46, 0], [14164, 1283.7, 160.60000000000002, 7.99, 0], [14165, 1672.0000000000002, 217.8, 7.68, 0], [14166, 1337.6000000000001, 147.4, 9.07, 0], [14167, 1577.4, 133.10000000000002, 11.85, 0], [14168, 4901.6, 531.3000000000001, 9.23, 0], [14169, 2801, 339, 8.26, 1], [14170, 3188, 368, 8.66, 1], [14171, 2367.2000000000003, 238.70000000000002, 9.92, 0], [14172, 1312.3000000000002, 135.3, 9.7, 0], [14173, 2611, 391, 6.68, 1], [14174, 1844.7, 196.9, 9.37, 0], [14175, 2424.4, 279.40000000000003, 8.68, 0], [14176, 1192.4, 136.4, 8.74, 0], [14177, 2851, 360, 7.92, 1], [14178, 2181.3, 218.9, 9.96, 0], [14179, 1857.9, 214.50000000000003, 8.66, 0], [14180, 2422.2000000000003, 298.1, 8.13, 0], [14181, 4085.4000000000005, 425.70000000000005, 9.6, 0], [14182, 3179.0000000000005, 277.20000000000005, 11.47, 0], [14183, 2578.4, 270.6, 9.53, 0], [14184, 2069.1000000000004, 214.50000000000003, 9.65, 0], [14185, 2315.5, 216.70000000000002, 10.69, 0], [14186, 1024.1000000000001, 118.80000000000001, 8.62, 0], [14187, 3639, 477, 7.63, 1], [14188, 3511, 379, 9.26, 1], [14190, 589, 95, 6.2, 1], [14191, 3008.5000000000005, 325.6, 9.24, 0], [14192, 2984.3, 322.3, 9.26, 0], [14193, 3093.2000000000003, 350.90000000000003, 8.82, 0], [14194, 2040.5000000000002, 213.4, 9.56, 0], [14195, 3491, 370, 9.44, 1], [14196, 2466.2000000000003, 239.8, 10.28, 0], [14197, 1934.9, 204.60000000000002, 9.46, 0], [14198, 1353, 164, 8.25, 1], [14199, 2051.5, 184.8, 11.1, 0], [14200, 3919, 438, 8.95, 1], [14201, 2135.1000000000004, 227.70000000000002, 9.38, 0], [14202, 2097.7000000000003, 243.10000000000002, 8.63, 0], [14203, 1519.1000000000001, 188.10000000000002, 8.08, 0], [14204, 2187.9, 287.1, 7.62, 0], [14205, 2752, 365, 7.54, 1], [14206, 1424.5000000000002, 157.3, 9.06, 0], [14207, 2820, 287, 9.83, 1], [14208, 4159, 528, 7.88, 1], [14209, 1455.3000000000002, 113.30000000000001, 12.84, 0], [14210, 3102, 347, 8.94, 1], [14211, 3534, 429, 8.24, 1], [14212, 2655, 350, 7.59, 1], [14213, 2213.2000000000003, 234.3, 9.45, 0], [14214, 2215.4, 233.20000000000002, 9.5, 0], [14215, 1840.3000000000002, 203.50000000000003, 9.04, 0], [14216, 1368.4, 130.9, 10.45, 0], [14217, 3590, 358, 10.03, 1], [14218, 3378.1000000000004, 324.5, 10.41, 0], [14219, 2949, 428, 6.89, 1], [14220, 3639, 369, 9.86, 1], [14221, 3765.3, 415.8, 9.06, 0], [14222, 2283.6000000000004, 247.50000000000003, 9.23, 0], [14223, 1403.6000000000001, 172.70000000000002, 8.13, 0], [14224, 1690.7, 199.10000000000002, 8.49, 0], [14225, 2060.3, 249.70000000000002, 8.25, 0], [14226, 1953.6000000000001, 272.8, 7.16, 0], [14227, 2618.0, 225.50000000000003, 11.61, 0], [14228, 2514.6000000000004, 337.70000000000005, 7.45, 0], [14229, 2117.5, 250.8, 8.44, 0], [14230, 2876.5000000000005, 305.8, 9.41, 0], [14231, 1931.6000000000001, 184.8, 10.45, 0], [14232, 2754.4, 209.00000000000003, 13.18, 0], [14233, 2192.3, 258.5, 8.48, 0], [14234, 3541, 375, 9.44, 1], [14235, 3845.6000000000004, 303.6, 12.67, 0], [14236, 3678.4, 405.90000000000003, 9.06, 0], [14237, 2932.6000000000004, 388.3, 7.55, 0], [14238, 2444.2000000000003, 247.50000000000003, 9.88, 0], [14240, 2767, 365, 7.58, 1], [14241, 1590.6000000000001, 165.0, 9.64, 0], [14242, 1556.5000000000002, 173.8, 8.96, 0], [14243, 2979, 362, 8.23, 1], [14244, 278.3, 4.4, 63.25, 0], [14245, 2823.7000000000003, 245.3, 11.51, 0], [14246, 377.3, 35.2, 10.72, 0], [14247, 3033.8, 288.20000000000005, 10.53, 0], [14248, 2150.5, 246.40000000000003, 8.73, 0], [14249, 2722, 353, 7.71, 1], [14250, 1994.3000000000002, 202.4, 9.85, 0], [14251, 2351.8, 269.5, 8.73, 0], [14252, 2134, 231, 9.24, 1], [14253, 2330.9, 220.00000000000003, 10.6, 0], [14254, 2151, 248, 8.67, 1], [14255, 1902, 257, 7.4, 1], [14256, 2938.1000000000004, 265.1, 11.08, 0], [14257, 2267.1000000000004, 262.90000000000003, 8.62, 0], [14258, 3538.7000000000003, 348.70000000000005, 10.15, 0], [14259, 1708.3000000000002, 172.70000000000002, 9.89, 0], [14260, 2525.6000000000004, 311.3, 8.11, 0], [14261, 2622.4, 259.6, 10.1, 0], [14262, 3507, 347, 10.11, 1], [14263, 3063.5000000000005, 394.90000000000003, 7.76, 0], [14264, 2042.7000000000003, 221.10000000000002, 9.24, 0], [14265, 1818.3000000000002, 212.3, 8.56, 0], [14266, 2464.0, 279.40000000000003, 8.82, 0], [14267, 3784.0000000000005, 348.70000000000005, 10.85, 0], [14268, 1926.1000000000001, 199.10000000000002, 9.67, 0], [14269, 2368.3, 204.60000000000002, 11.58, 0], [14270, 5093.0, 344.3, 14.79, 0], [14271, 2194.5, 250.8, 8.75, 0], [14272, 1377.2, 180.4, 7.63, 0], [14273, 1058.2, 103.4, 10.23, 0], [14274, 2273.7000000000003, 258.5, 8.8, 0], [14275, 2197.8, 236.50000000000003, 9.29, 0], [14276, 1982.2000000000003, 244.20000000000002, 8.12, 0], [14277, 2041.6000000000001, 206.8, 9.87, 0], [14278, 1741.3000000000002, 205.70000000000002, 8.47, 0], [14279, 3860, 373, 10.35, 1], [14280, 1993.2000000000003, 237.60000000000002, 8.39, 0], [14281, 2801.7000000000003, 284.90000000000003, 9.83, 0], [14282, 2526.7000000000003, 260.70000000000005, 9.69, 0], [14283, 1949.2, 270.6, 7.2, 0], [14284, 1906.3000000000002, 195.8, 9.74, 0], [14285, 2163.7000000000003, 237.60000000000002, 9.11, 0], [14286, 2325.4, 237.60000000000002, 9.79, 0], [14287, 1519.1000000000001, 217.8, 6.97, 0], [14288, 2931.5000000000005, 259.6, 11.29, 0], [14289, 1129.7, 125.4, 9.01, 0], [14290, 1786.4, 193.60000000000002, 9.23, 0], [14291, 2227.5, 270.6, 8.23, 0], [14292, 724.9000000000001, 95.7, 7.57, 0], [14293, 2074.6000000000004, 231.00000000000003, 8.98, 0], [14294, 2314.4, 303.6, 7.62, 0], [14295, 2856.7000000000003, 288.20000000000005, 9.91, 0], [14296, 3059.1000000000004, 309.1, 9.9, 0], [14297, 3125.1000000000004, 337.70000000000005, 9.25, 0], [14298, 2480.5, 336.6, 7.37, 0], [14299, 1942.6000000000001, 174.9, 11.11, 0], [14300, 2159.3, 239.8, 9.0, 0], [14301, 3275.8, 343.20000000000005, 9.54, 0], [14302, 2684.0, 268.40000000000003, 10.0, 0], [14303, 2423.3, 271.70000000000005, 8.92, 0], [14304, 3100, 376, 8.24, 1], [14305, 4580.400000000001, 368.50000000000006, 12.43, 0], [14306, 1951.4, 284.90000000000003, 6.85, 0], [14307, 3352.8, 297.0, 11.29, 0], [14308, 2025.1000000000001, 218.9, 9.25, 0], [14309, 3287.9, 358.6, 9.17, 0], [14310, 4033.7000000000003, 320.1, 12.6, 0], [14311, 1458.6000000000001, 189.20000000000002, 7.71, 0], [14312, 2325.4, 261.8, 8.88, 0], [14313, 3438.6000000000004, 315.70000000000005, 10.89, 0], [14314, 2147.2000000000003, 188.10000000000002, 11.42, 0], [14315, 3397.9, 287.1, 11.84, 0], [14316, 2161.5, 234.3, 9.23, 0], [14317, 2231.9, 239.8, 9.31, 0], [14318, 1976.7000000000003, 199.10000000000002, 9.93, 0], [14319, 2207.7000000000003, 266.20000000000005, 8.29, 0], [14320, 3170.2000000000003, 226.60000000000002, 13.99, 0], [14321, 2942, 385, 7.64, 1], [14322, 2202.2000000000003, 223.3, 9.86, 0], [14323, 2511, 360, 6.98, 1], [14324, 2619.1000000000004, 224.4, 11.67, 0], [14325, 2022.9, 229.9, 8.8, 0], [14326, 1387.1000000000001, 154.0, 9.01, 0], [14327, 3052, 395, 7.73, 1], [14329, 1356.3000000000002, 133.10000000000002, 10.19, 0], [14330, 910.8000000000001, 135.3, 6.73, 0], [14331, 2051.5, 200.20000000000002, 10.25, 0], [14332, 2264.9, 224.4, 10.09, 0], [14333, 2549.8, 294.8, 8.65, 0], [14334, 2692.8, 291.5, 9.24, 0], [14335, 1785.3000000000002, 217.8, 8.2, 0], [14336, 2825.9, 328.90000000000003, 8.59, 0], [14337, 3633, 401, 9.06, 1], [14338, 1650.0000000000002, 158.4, 10.42, 0], [14339, 1510.3000000000002, 178.20000000000002, 8.48, 0], [14340, 2006.4, 221.10000000000002, 9.07, 0], [14341, 1039.5, 104.50000000000001, 9.95, 0], [14342, 3208, 426, 7.53, 1], [14343, 2538, 313, 8.11, 1], [14344, 900, 111, 8.11, 1], [14345, 2753.3, 249.70000000000002, 11.03, 0], [14346, 1722.6000000000001, 170.5, 10.1, 0], [14347, 2439.8, 273.90000000000003, 8.91, 0], [14348, 1879.9, 222.20000000000002, 8.46, 0], [14349, 2913.9, 283.8, 10.27, 0], [14350, 3041, 388, 7.84, 1], [14351, 2996, 402, 7.45, 1], [14352, 3306, 412, 8.02, 1], [14353, 1375.0, 182.60000000000002, 7.53, 0], [14354, 3340.7000000000003, 341.0, 9.8, 0], [14355, 1901.9, 205.70000000000002, 9.25, 0], [14356, 2955.7000000000003, 298.1, 9.92, 0], [14357, 2171.4, 233.20000000000002, 9.31, 0], [14358, 2059.2000000000003, 270.6, 7.61, 0], [14359, 1907.4, 222.20000000000002, 8.58, 0], [14360, 3069.0000000000005, 349.8, 8.77, 0], [14361, 1585.1000000000001, 201.3, 7.87, 0], [14362, 3225, 393, 8.21, 1], [14363, 9.9, 1.1, 9.0, 0], [14364, 2542, 228, 11.15, 1], [14365, 1186.9, 160.60000000000002, 7.39, 0], [14366, 2794.0, 262.90000000000003, 10.63, 0], [14367, 2910.6000000000004, 255.20000000000002, 11.41, 0], [14368, 2052.6000000000004, 229.9, 8.93, 0], [14370, 3490.3, 335.5, 10.4, 0], [14371, 2129.6000000000004, 250.8, 8.49, 0], [14372, 1019.7, 108.9, 9.36, 0], [14373, 1507.0000000000002, 179.3, 8.4, 0], [14374, 1347.5, 183.70000000000002, 7.34, 0], [14375, 3457, 390, 8.86, 1], [14376, 2415.6000000000004, 192.50000000000003, 12.55, 0], [14377, 2458.5, 256.3, 9.59, 0], [14378, 2131.8, 275.0, 7.75, 0], [14380, 1060.4, 115.50000000000001, 9.18, 0], [14381, 1289.2, 193.60000000000002, 6.66, 0], [14383, 1365.1000000000001, 199.10000000000002, 6.86, 0], [14384, 2227.5, 438.90000000000003, 5.08, 0], [14385, 1656.6000000000001, 184.8, 8.96, 0], [14386, 1343.1000000000001, 143.0, 9.39, 0], [14387, 1058.2, 145.20000000000002, 7.29, 0], [14388, 1246.3000000000002, 161.70000000000002, 7.71, 0], [14389, 1348.6000000000001, 182.60000000000002, 7.39, 0], [14390, 1215.5, 174.9, 6.95, 0], [14391, 1501.5000000000002, 180.4, 8.32, 0], [14392, 1455.3000000000002, 173.8, 8.37, 0], [14393, 1980.0000000000002, 183.70000000000002, 10.78, 0], [14394, 1442.1000000000001, 158.4, 9.1, 0], [14395, 1128.6000000000001, 171.60000000000002, 6.58, 0], [14396, 1999.8000000000002, 245.3, 8.15, 0], [14397, 1050.5, 159.5, 6.59, 0], [14398, 1347.5, 113.30000000000001, 11.89, 0], [14399, 1189.1000000000001, 148.5, 8.01, 0], [14400, 943.8000000000001, 117.7, 8.02, 0], [14401, 2080.1000000000004, 171.60000000000002, 12.12, 0], [14402, 1513.6000000000001, 158.4, 9.56, 0], [14403, 1787, 167, 10.7, 1], [14404, 1300.2, 198.00000000000003, 6.57, 0], [14405, 1757.8000000000002, 143.0, 12.29, 0], [14406, 963.6, 112.2, 8.59, 0], [14407, 2570.7000000000003, 207.9, 12.37, 0], [14408, 2810, 281, 10.0, 1], [14410, 2183.5, 319.0, 6.84, 0], [14411, 1236.4, 182.60000000000002, 6.77, 0], [14412, 2558.6000000000004, 458.70000000000005, 5.58, 0], [14413, 1861.2, 133.10000000000002, 13.98, 0], [14414, 783.2, 70.4, 11.12, 0], [14415, 1843.6000000000001, 194.70000000000002, 9.47, 0], [14416, 2054.8, 326.70000000000005, 6.29, 0], [14417, 2136.2000000000003, 268.40000000000003, 7.96, 0], [14418, 1718.2, 174.9, 9.82, 0], [14421, 2887, 348, 8.3, 1], [14422, 866.8000000000001, 89.10000000000001, 9.73, 0], [14423, 1474.0000000000002, 203.50000000000003, 7.24, 0], [14424, 1648.9, 176.0, 9.37, 0], [14425, 1205.6000000000001, 158.4, 7.61, 0], [14426, 1709.4, 212.3, 8.05, 0], [14427, 1205.6000000000001, 189.20000000000002, 6.37, 0], [14428, 1108.8000000000002, 112.2, 9.88, 0], [14429, 2238.5, 221.10000000000002, 10.12, 0], [14430, 2092.2000000000003, 222.20000000000002, 9.42, 0], [14431, 3143.8, 267.3, 11.76, 0], [14432, 1325.5, 128.70000000000002, 10.3, 0], [14433, 2121.9, 216.70000000000002, 9.79, 0], [14434, 2345, 343, 6.84, 1], [14435, 1839.2, 213.4, 8.62, 0], [14436, 1035.1000000000001, 125.4, 8.25, 0], [14437, 2625.7000000000003, 489.50000000000006, 5.36, 0], [14438, 1688.5000000000002, 161.70000000000002, 10.44, 0], [14439, 1128.6000000000001, 185.9, 6.07, 0], [14440, 1685.2, 150.70000000000002, 11.18, 0], [14441, 3192.2000000000003, 507.1, 6.3, 0], [14442, 887.7, 125.4, 7.08, 0], [14443, 3010.7000000000003, 425.70000000000005, 7.07, 0], [14444, 1934.9, 177.10000000000002, 10.93, 0], [14445, 2169.2000000000003, 209.00000000000003, 10.38, 0], [14446, 1601.6000000000001, 201.3, 7.96, 0], [14447, 1874.4, 226.60000000000002, 8.27, 0], [14448, 2026.2000000000003, 178.20000000000002, 11.37, 0], [14449, 2976, 378, 7.87, 1], [14450, 2823.7000000000003, 250.8, 11.26, 0], [14454, 1276.0, 126.50000000000001, 10.09, 0], [14455, 3377.0000000000005, 311.3, 10.85, 0], [14456, 2836.9, 272.8, 10.4, 0], [14457, 2108.7000000000003, 206.8, 10.2, 0], [14458, 2726.9, 261.8, 10.42, 0], [14459, 2161, 275, 7.86, 1], [14460, 2227.5, 245.3, 9.08, 0], [14461, 2197.8, 215.60000000000002, 10.19, 0], [14462, 1868.9, 184.8, 10.11, 0], [14463, 1683.0000000000002, 204.60000000000002, 8.23, 0], [14464, 1692.9, 179.3, 9.44, 0], [14466, 2303.4, 183.70000000000002, 12.54, 0], [14467, 2318.8, 232.10000000000002, 9.99, 0], [14468, 1409.1000000000001, 183.70000000000002, 7.67, 0], [14469, 2429.9, 281.6, 8.63, 0], [14470, 2297.9, 251.90000000000003, 9.12, 0], [14471, 2002.0000000000002, 210.10000000000002, 9.53, 0], [14472, 2798.4, 305.8, 9.15, 0], [14473, 2180.2000000000003, 232.10000000000002, 9.39, 0], [14474, 1789.7, 209.00000000000003, 8.56, 0], [14475, 2680.7000000000003, 240.9, 11.13, 0], [14476, 2860, 371, 7.71, 1], [14477, 1983.3000000000002, 210.10000000000002, 9.44, 0], [14478, 2565, 328, 7.82, 1], [14479, 2446, 267, 9.16, 1], [14480, 2503.6000000000004, 259.6, 9.64, 0], [14481, 3411, 371, 9.19, 1], [14482, 2729.1000000000004, 275.0, 9.92, 0], [14483, 2750, 295, 9.32, 1], [14485, 2435.4, 203.50000000000003, 11.97, 0], [14486, 3414, 422, 8.09, 1], [14487, 2765.4, 270.6, 10.22, 0], [14488, 2660.9, 346.5, 7.68, 0], [14489, 2279.2000000000003, 210.10000000000002, 10.85, 0], [14491, 2644, 290, 9.12, 1], [14492, 2092.2000000000003, 204.60000000000002, 10.23, 0], [14493, 3017, 399, 7.56, 1], [14494, 3000.8, 292.6, 10.26, 0], [14495, 2585.0, 243.10000000000002, 10.63, 0], [14496, 3107.5000000000005, 348.70000000000005, 8.91, 0], [14497, 2800.6000000000004, 324.5, 8.63, 0], [14498, 3446, 431, 8.0, 1], [14499, 2535, 286, 8.86, 1], [14500, 3131.7000000000003, 290.40000000000003, 10.78, 0], [14501, 2018, 286, 7.06, 1], [14502, 3060, 301, 10.17, 1], [14503, 2101.0, 237.60000000000002, 8.84, 0], [14504, 2774.2000000000003, 295.90000000000003, 9.38, 0], [14505, 2120.8, 218.9, 9.69, 0], [14506, 2286.9, 242.00000000000003, 9.45, 0], [14507, 1875.5000000000002, 211.20000000000002, 8.88, 0], [14508, 2757, 424, 6.5, 1], [14510, 1720.4, 174.9, 9.84, 0], [14511, 1874.4, 221.10000000000002, 8.48, 0], [14512, 2776.4, 289.3, 9.6, 0], [14513, 2890.8, 303.6, 9.52, 0], [14514, 1930.5000000000002, 253.00000000000003, 7.63, 0], [14515, 1663.2, 203.50000000000003, 8.17, 0], [14516, 2481.6000000000004, 315.70000000000005, 7.86, 0], [14517, 3558, 425, 8.37, 1], [14518, 2470.6000000000004, 270.6, 9.13, 0], [14519, 2110.9, 237.60000000000002, 8.88, 0], [14521, 1895.3000000000002, 255.20000000000002, 7.43, 0], [14522, 1872.2, 253.00000000000003, 7.4, 0], [14523, 2073.5, 269.5, 7.69, 0], [14524, 2780.8, 354.20000000000005, 7.85, 0], [14525, 2013.0000000000002, 225.50000000000003, 8.93, 0], [14526, 2092.2000000000003, 246.40000000000003, 8.49, 0], [14527, 2294.6000000000004, 277.20000000000005, 8.28, 0], [14528, 1764, 337, 5.23, 1], [14529, 2670.8, 268.40000000000003, 9.95, 0], [14530, 3111.9, 322.3, 9.66, 0], [14531, 3118.5000000000005, 436.70000000000005, 7.14, 0], [14532, 1881.0000000000002, 246.40000000000003, 7.63, 0], [14533, 2684.0, 262.90000000000003, 10.21, 0], [14534, 2141.7000000000003, 214.50000000000003, 9.98, 0], [14535, 1366.2, 185.9, 7.35, 0], [14536, 3004, 393, 7.64, 1], [14537, 2204.4, 270.6, 8.15, 0], [14538, 3437, 423, 8.13, 1], [14539, 2080.1000000000004, 223.3, 9.32, 0], [14540, 2044.9, 212.3, 9.63, 0], [14541, 2264, 304, 7.45, 1], [14542, 4677, 375, 12.47, 1], [14543, 3148, 462, 6.81, 1], [14544, 2862.2000000000003, 335.5, 8.53, 0], [14545, 2414.5, 243.10000000000002, 9.93, 0], [14546, 11.0, 1.1, 10.0, 0], [14547, 3283.5000000000005, 314.6, 10.44, 0], [14548, 3259.3, 376.20000000000005, 8.66, 0], [14549, 1192.4, 133.10000000000002, 8.96, 0], [14550, 2098.8, 216.70000000000002, 9.69, 0], [14551, 2182, 314, 6.95, 1], [14552, 2588, 320, 8.09, 1], [14553, 1725.9, 203.50000000000003, 8.48, 0], [14554, 2866.6000000000004, 271.70000000000005, 10.55, 0], [14555, 2206.6000000000004, 265.1, 8.32, 0], [14556, 1776.5000000000002, 201.3, 8.83, 0], [14557, 2257.2000000000003, 240.9, 9.37, 0], [14558, 2035.0000000000002, 272.8, 7.46, 0], [14559, 2544, 346, 7.35, 1], [14560, 1756.7, 213.4, 8.23, 0], [14561, 137.5, 15.400000000000002, 8.93, 0], [14562, 1758, 278, 6.32, 1], [14563, 4181.1, 305.8, 13.67, 0], [14564, 510.40000000000003, 39.6, 12.89, 0], [14565, 2765.4, 246.40000000000003, 11.22, 0], [14567, 2006.4, 218.9, 9.17, 0], [14568, 1932, 300, 6.44, 1], [14569, 2718.1000000000004, 279.40000000000003, 9.73, 0], [14570, 2203, 329, 6.7, 1], [14571, 2425, 330, 7.35, 1], [14572, 2249.5, 238.70000000000002, 9.42, 0], [14573, 3324, 377, 8.82, 1], [14574, 2865, 398, 7.2, 1], [14575, 2740, 415, 6.6, 1], [14576, 3509, 438, 8.01, 1], [14577, 149.60000000000002, 9.9, 15.11, 0], [14578, 2743, 392, 7.0, 1], [14579, 1128.6000000000001, 123.20000000000002, 9.16, 0], [14580, 1323.3000000000002, 149.60000000000002, 8.85, 0], [14581, 2020.7000000000003, 270.6, 7.47, 0], [14582, 2981.0000000000005, 334.40000000000003, 8.91, 0], [14583, 2960.1000000000004, 306.90000000000003, 9.65, 0], [14584, 2431.0, 268.40000000000003, 9.06, 0], [14585, 2204.4, 265.1, 8.32, 0], [14586, 2977, 328, 9.08, 1], [14587, 2019.6000000000001, 255.20000000000002, 7.91, 0], [14588, 3898, 430, 9.07, 1], [14589, 2939.2000000000003, 309.1, 9.51, 0], [14590, 4185, 501, 8.35, 1], [14591, 2426.6000000000004, 181.50000000000003, 13.37, 0], [14592, 2349.6000000000004, 248.60000000000002, 9.45, 0], [14593, 2509.1000000000004, 310.20000000000005, 8.09, 0], [14594, 1956, 260, 7.52, 1], [14595, 2409, 346, 6.96, 1], [14596, 2777, 413, 6.72, 1], [14597, 1966.8000000000002, 199.10000000000002, 9.88, 0], [14598, 498.30000000000007, 47.300000000000004, 10.53, 0], [14599, 1668.7, 217.8, 7.66, 0], [14600, 2458, 326, 7.54, 1], [14601, 4225.1, 473.00000000000006, 8.93, 0], [14602, 1867.8000000000002, 240.9, 7.75, 0], [14603, 3241, 363, 8.93, 1], [14604, 4290.0, 409.20000000000005, 10.48, 0], [14605, 2011.9, 275.0, 7.32, 0], [14606, 1823.8000000000002, 238.70000000000002, 7.64, 0], [14607, 2525.6000000000004, 308.0, 8.2, 0], [14608, 1986.6000000000001, 199.10000000000002, 9.98, 0], [14609, 1884.3000000000002, 196.9, 9.57, 0], [14610, 2525.6000000000004, 260.70000000000005, 9.69, 0], [14611, 2206, 315, 7.0, 1], [14612, 2544.3, 267.3, 9.52, 0], [14613, 3284, 368, 8.92, 1], [14614, 1471.8000000000002, 147.4, 9.99, 0], [14615, 2691.7000000000003, 302.5, 8.9, 0], [14616, 1882.1000000000001, 212.3, 8.87, 0], [14617, 2196.7000000000003, 254.10000000000002, 8.65, 0], [14618, 2148.3, 249.70000000000002, 8.6, 0], [14619, 2429.9, 264.0, 9.2, 0], [14620, 2961.2000000000003, 338.8, 8.74, 0], [14621, 2790.7000000000003, 372.90000000000003, 7.48, 0], [14622, 2476.1000000000004, 292.6, 8.46, 0], [14623, 1806.2, 183.70000000000002, 9.83, 0], [14624, 1678.6000000000001, 207.9, 8.07, 0], [14625, 3906.1000000000004, 422.40000000000003, 9.25, 0], [14626, 3323.1000000000004, 355.3, 9.35, 0], [14627, 1820.5000000000002, 141.9, 12.83, 0], [14628, 2180.2000000000003, 265.1, 8.22, 0], [14629, 3099.8, 268.40000000000003, 11.55, 0], [14630, 3608, 489, 7.38, 1], [14631, 1290.3000000000002, 116.60000000000001, 11.07, 0], [14632, 2894.1000000000004, 322.3, 8.98, 0], [14633, 2024.0000000000002, 254.10000000000002, 7.97, 0], [14634, 4636.5, 377.3, 12.29, 0], [14635, 3320.9, 381.70000000000005, 8.7, 0], [14636, 2712.6000000000004, 324.5, 8.36, 0], [14637, 1988.8000000000002, 244.20000000000002, 8.14, 0], [14638, 1752.3000000000002, 194.70000000000002, 9.0, 0], [14639, 2700.5, 308.0, 8.77, 0], [14640, 1689.6000000000001, 162.8, 10.38, 0], [14641, 2832.5000000000005, 255.20000000000002, 11.1, 0], [14643, 3333.0000000000005, 366.3, 9.1, 0], [14644, 2757.7000000000003, 281.6, 9.79, 0], [14645, 2480.5, 294.8, 8.41, 0], [14646, 1675.3000000000002, 229.9, 7.29, 0], [14647, 3028.3, 325.6, 9.3, 0], [14648, 3374.8, 374.00000000000006, 9.02, 0], [14649, 3349.5000000000005, 410.3, 8.16, 0], [14650, 2552.0, 315.70000000000005, 8.08, 0], [14651, 2365.0, 246.40000000000003, 9.6, 0], [14652, 2568.5, 257.40000000000003, 9.98, 0], [14653, 3701.5000000000005, 386.1, 9.59, 0], [14654, 2253.9, 258.5, 8.72, 0], [14655, 2179.1000000000004, 297.0, 7.34, 0], [14656, 2208.8, 215.60000000000002, 10.24, 0], [14657, 2643.3, 272.8, 9.69, 0], [14658, 2359.5, 237.60000000000002, 9.93, 0], [14659, 2134.0, 204.60000000000002, 10.43, 0], [14660, 1477.3000000000002, 192.50000000000003, 7.67, 0], [14661, 2251.7000000000003, 235.4, 9.57, 0], [14662, 2708.2000000000003, 256.3, 10.57, 0], [14663, 1585.1000000000001, 185.9, 8.53, 0], [14664, 1923.9, 232.10000000000002, 8.29, 0], [14665, 3360, 425, 7.91, 1], [14666, 3233, 291, 11.11, 1], [14667, 2856.7000000000003, 326.70000000000005, 8.74, 0], [14668, 2126.3, 254.10000000000002, 8.37, 0], [14669, 2590.5, 294.8, 8.79, 0], [14670, 1890.9, 217.8, 8.68, 0], [14671, 4448.400000000001, 431.20000000000005, 10.32, 0], [14672, 1769.9, 167.20000000000002, 10.59, 0], [14673, 3431, 409, 8.39, 1], [14674, 2447.5, 294.8, 8.3, 0], [14675, 2010.8000000000002, 217.8, 9.23, 0], [14676, 2396.9, 244.20000000000002, 9.82, 0], [14677, 2589.4, 339.90000000000003, 7.62, 0], [14678, 4359.3, 476.3, 9.15, 0], [14679, 2455.2000000000003, 246.40000000000003, 9.96, 0], [14680, 2071.3, 262.90000000000003, 7.88, 0], [14681, 2810.5, 246.40000000000003, 11.41, 0], [14682, 3856.6000000000004, 402.6, 9.58, 0], [14683, 3352, 413, 8.12, 1], [14684, 3073.4, 321.20000000000005, 9.57, 0], [14685, 2957, 380, 7.78, 1], [14686, 3493, 398, 8.78, 1], [14687, 2415.6000000000004, 227.70000000000002, 10.61, 0], [14688, 2377.1000000000004, 297.0, 8.0, 0], [14689, 1807.3000000000002, 196.9, 9.18, 0], [14690, 2145.0, 190.3, 11.27, 0], [14691, 1832.6000000000001, 207.9, 8.81, 0], [14692, 2327.6000000000004, 306.90000000000003, 7.58, 0], [14693, 3749, 484, 7.75, 1], [14694, 2413, 344, 7.01, 1], [14695, 2165, 324, 6.68, 1], [14696, 3226, 455, 7.09, 1], [14697, 2645, 381, 6.94, 1], [14698, 3093.2000000000003, 376.20000000000005, 8.22, 0], [14699, 3253, 385, 8.45, 1], [14700, 4032.6000000000004, 372.90000000000003, 10.81, 0], [14701, 2813, 388, 7.25, 1], [14702, 4400, 330, 13.33, 1], [14703, 2284.7000000000003, 278.3, 8.21, 0], [14704, 2145.0, 244.20000000000002, 8.78, 0], [14705, 1714.9, 209.00000000000003, 8.21, 0], [14706, 2324.3, 287.1, 8.1, 0], [14707, 3052.5000000000005, 365.20000000000005, 8.36, 0], [14708, 2874.3, 292.6, 9.82, 0], [14709, 3005.2000000000003, 322.3, 9.32, 0], [14710, 1130.8000000000002, 127.60000000000001, 8.86, 0], [14711, 2632, 401, 6.56, 1], [14712, 3225, 443, 7.28, 1], [14713, 3289, 264, 12.46, 1], [14714, 2857, 305, 9.37, 1], [14715, 3173, 400, 7.93, 1], [14716, 51.7, 7.700000000000001, 6.71, 0], [14717, 2549.8, 317.90000000000003, 8.02, 0], [14718, 2710, 394, 6.88, 1], [14719, 2525.6000000000004, 326.70000000000005, 7.73, 0], [14721, 2294.6000000000004, 244.20000000000002, 9.4, 0], [14722, 1808.4, 156.20000000000002, 11.58, 0], [14723, 3938.0000000000005, 441.1, 8.93, 0], [14724, 1828.2, 212.3, 8.61, 0], [14725, 2547, 340, 7.49, 1], [14726, 1453.1000000000001, 138.60000000000002, 10.48, 0], [14727, 2139.5, 235.4, 9.09, 0], [14728, 1063.7, 130.9, 8.13, 0], [14729, 786.5000000000001, 89.10000000000001, 8.83, 0], [14730, 1740.2, 187.00000000000003, 9.31, 0], [14731, 1773.2, 162.8, 10.89, 0], [14732, 1148.4, 150.70000000000002, 7.62, 0], [14733, 2642.2000000000003, 302.5, 8.73, 0], [14734, 1776.5000000000002, 284.90000000000003, 6.24, 0], [14735, 1283.7, 141.9, 9.05, 0], [14736, 4724, 349, 13.54, 1], [14737, 2153.8, 249.70000000000002, 8.63, 0], [14738, 2276, 287, 7.93, 1], [14739, 1812.8000000000002, 199.10000000000002, 9.1, 0], [14740, 2693.9, 256.3, 10.51, 0], [14741, 2638.9, 294.8, 8.95, 0], [14742, 1789.7, 215.60000000000002, 8.3, 0], [14743, 1826.0000000000002, 216.70000000000002, 8.43, 0], [14744, 1815.0000000000002, 224.4, 8.09, 0], [14745, 1859.0000000000002, 227.70000000000002, 8.16, 0], [14746, 1898.6000000000001, 198.00000000000003, 9.59, 0], [14747, 2825.9, 312.40000000000003, 9.05, 0], [14748, 1861.2, 213.4, 8.72, 0], [14749, 2623.5, 304.70000000000005, 8.61, 0], [14750, 2267.1000000000004, 314.6, 7.21, 0], [14751, 1615.9, 218.9, 7.38, 0], [14752, 2669.7000000000003, 265.1, 10.07, 0], [14753, 2175.8, 222.20000000000002, 9.79, 0], [14754, 2448.6000000000004, 277.20000000000005, 8.83, 0], [14755, 2763.2000000000003, 272.8, 10.13, 0], [14756, 2084.5, 236.50000000000003, 8.81, 0], [14757, 2728.0, 332.20000000000005, 8.21, 0], [14758, 3660.8, 419.1, 8.73, 0], [14759, 2272.6000000000004, 220.00000000000003, 10.33, 0], [14760, 1853.5000000000002, 199.10000000000002, 9.31, 0], [14761, 2924.9, 316.8, 9.23, 0], [14762, 1960.2, 220.00000000000003, 8.91, 0], [14763, 2688.4, 270.6, 9.93, 0], [14764, 3436.4, 400.40000000000003, 8.58, 0], [14765, 2374.9, 323.40000000000003, 7.34, 0], [14766, 1937.1000000000001, 249.70000000000002, 7.76, 0], [14767, 2176.9, 233.20000000000002, 9.33, 0], [14768, 1724.8000000000002, 211.20000000000002, 8.17, 0], [14769, 1910.7, 223.3, 8.56, 0], [14770, 2218.7000000000003, 291.5, 7.61, 0], [14771, 3788, 350, 10.82, 1], [14772, 2601.5, 354.20000000000005, 7.34, 0], [14773, 2335.3, 237.60000000000002, 9.83, 0], [14774, 2741.2000000000003, 281.6, 9.73, 0], [14775, 2253.9, 196.9, 11.45, 0], [14776, 2599.3, 249.70000000000002, 10.41, 0], [14777, 2544.3, 223.3, 11.39, 0], [14778, 1705.0000000000002, 243.10000000000002, 7.01, 0], [14779, 3414.4, 408.1, 8.37, 0], [14780, 2664.2000000000003, 295.90000000000003, 9.0, 0], [14781, 1424.5000000000002, 125.4, 11.36, 0], [14782, 2246.2000000000003, 273.90000000000003, 8.2, 0], [14783, 3052.5000000000005, 352.0, 8.67, 0], [14784, 1399.2, 223.3, 6.27, 0], [14785, 2987.6000000000004, 290.40000000000003, 10.29, 0], [14786, 2715.9, 316.8, 8.57, 0], [14787, 1686.3000000000002, 232.10000000000002, 7.27, 0], [14788, 2185.7000000000003, 236.50000000000003, 9.24, 0], [14789, 3145, 360, 8.74, 1], [14790, 3041.5000000000005, 352.0, 8.64, 0], [14791, 2745.6000000000004, 248.60000000000002, 11.04, 0], [14792, 2181.3, 237.60000000000002, 9.18, 0], [14793, 1684.1000000000001, 169.4, 9.94, 0], [14794, 3765, 363, 10.37, 1], [14795, 1163.8000000000002, 117.7, 9.89, 0], [14796, 2708.2000000000003, 283.8, 9.54, 0], [14797, 2327.6000000000004, 319.0, 7.3, 0], [14798, 1804.0000000000002, 202.4, 8.91, 0], [14799, 1386.0, 135.3, 10.24, 0], [14800, 3102.0000000000005, 320.1, 9.69, 0], [14801, 2546.5, 272.8, 9.33, 0], [14803, 2401.3, 284.90000000000003, 8.43, 0], [14804, 2568, 342, 7.51, 1], [14805, 3714, 402, 9.24, 1], [14806, 1615.9, 147.4, 10.96, 0], [14807, 1931.6000000000001, 166.10000000000002, 11.63, 0], [14808, 3131, 382, 8.2, 1], [14809, 2161.5, 225.50000000000003, 9.59, 0], [14810, 3594.8, 396.00000000000006, 9.08, 0], [14811, 2823, 324, 8.71, 1], [14812, 3096.5000000000005, 237.60000000000002, 13.03, 0], [14813, 2599.3, 271.70000000000005, 9.57, 0], [14814, 2118.6000000000004, 290.40000000000003, 7.3, 0], [14815, 2513.5, 297.0, 8.46, 0], [14816, 2350.7000000000003, 242.00000000000003, 9.71, 0], [14817, 2251.7000000000003, 304.70000000000005, 7.39, 0], [14818, 2347.4, 277.20000000000005, 8.47, 0], [14819, 2018.5000000000002, 239.8, 8.42, 0], [14820, 2213.2000000000003, 290.40000000000003, 7.62, 0], [14821, 957.0000000000001, 117.7, 8.13, 0], [14822, 1917.3000000000002, 226.60000000000002, 8.46, 0], [14823, 3071.2000000000003, 310.20000000000005, 9.9, 0], [14824, 2207, 313, 7.05, 1], [14825, 3377, 293, 11.53, 1], [14826, 2358.4, 297.0, 7.94, 0], [14827, 2605, 289, 9.01, 1], [14828, 1903.0000000000002, 199.10000000000002, 9.56, 0], [14829, 2550.9, 268.40000000000003, 9.5, 0], [14830, 2497.0, 324.5, 7.69, 0], [14831, 1783.1000000000001, 181.50000000000003, 9.82, 0], [14833, 2502.5, 297.0, 8.43, 0], [14834, 4177, 347, 12.04, 1], [14835, 2013.0000000000002, 226.60000000000002, 8.88, 0], [14836, 2222, 314, 7.08, 1], [14837, 2692, 327, 8.23, 1], [14838, 1961.3000000000002, 261.8, 7.49, 0], [14839, 1774.3000000000002, 213.4, 8.31, 0], [14840, 4236.1, 378.40000000000003, 11.19, 0], [14841, 1576.3000000000002, 231.00000000000003, 6.82, 0], [14842, 2977.7000000000003, 375.1, 7.94, 0], [14843, 2633, 376, 7.0, 1], [14844, 925, 219, 4.22, 1], [14845, 3036, 317, 9.58, 1], [14846, 2726, 358, 7.61, 1], [14847, 280.5, 15.400000000000002, 18.21, 0], [14848, 3724.6000000000004, 422.40000000000003, 8.82, 0], [14849, 2550.9, 260.70000000000005, 9.78, 0], [14850, 3497, 440, 7.95, 1], [14851, 2702, 342, 7.9, 1], [14852, 2115.3, 279.40000000000003, 7.57, 0], [14853, 2985, 374, 7.98, 1], [14854, 2662, 363, 7.33, 1], [14855, 3449, 390, 8.84, 1], [14856, 4473, 382, 11.71, 1], [14857, 2840.2000000000003, 291.5, 9.74, 0], [14858, 2027.3000000000002, 278.3, 7.28, 0], [14859, 3252.7000000000003, 366.3, 8.88, 0], [14860, 2184, 284, 7.69, 1], [14861, 1853.5000000000002, 207.9, 8.92, 0], [14862, 1776, 294, 6.04, 1], [14863, 3085.5000000000005, 275.0, 11.22, 0], [14864, 1268.3000000000002, 137.5, 9.22, 0], [14865, 2449.7000000000003, 300.3, 8.16, 0], [14866, 3224, 384, 8.4, 1], [14867, 2846, 357, 7.97, 1], [14868, 4736, 465, 10.18, 1], [14869, 2807, 305, 9.2, 1], [14870, 2149.4, 233.20000000000002, 9.22, 0], [14871, 2859, 337, 8.48, 1], [14872, 992.2, 111.10000000000001, 8.93, 0], [14873, 1977.8000000000002, 240.9, 8.21, 0], [14874, 1541.1000000000001, 145.20000000000002, 10.61, 0], [14875, 3895, 450, 8.66, 1], [14876, 1102.2, 137.5, 8.02, 0], [14877, 2444, 349, 7.0, 1], [14878, 2268, 317, 7.15, 1], [14879, 1598.3000000000002, 181.50000000000003, 8.81, 0], [14880, 2978.8, 280.5, 10.62, 0], [14881, 3298, 396, 8.33, 1], [14882, 2653.2000000000003, 305.8, 8.68, 0], [14883, 3961.1000000000004, 401.50000000000006, 9.87, 0], [14884, 168, 27, 6.22, 1], [14885, 666, 64, 10.41, 1], [14886, 1286, 131, 9.82, 1], [14887, 866, 117, 7.4, 1], [14888, 589, 65, 9.06, 1], [14889, 1661, 154, 10.79, 1], [14890, 1200, 189, 6.35, 1], [14891, 939, 159, 5.91, 1], [14892, 32, 4, 8.0, 1], [14893, 1024, 90, 11.38, 1], [14894, 1293, 124, 10.43, 1], [14895, 1148, 130, 8.83, 1], [14896, 954, 90, 10.6, 1], [14897, 2173, 227, 9.57, 1], [14898, 982, 130, 7.55, 1], [14899, 959, 151, 6.35, 1], [14900, 1789, 149, 12.01, 1], [14901, 685, 85, 8.06, 1], [14902, 1040, 120, 8.67, 1], [14903, 90, 9, 10.0, 1], [14904, 1005, 94, 10.69, 1], [14905, 626, 78, 8.03, 1], [14906, 1067, 110, 9.7, 1], [14907, 843, 95, 8.87, 1], [14908, 2516, 277, 9.08, 1], [14910, 638, 92, 6.93, 1], [14911, 701, 77, 9.1, 1], [14912, 496, 68, 7.29, 1], [14913, 386, 95, 4.06, 1], [14914, 3102, 374, 8.29, 1], [14915, 1247, 138, 9.04, 1], [14916, 1296, 168, 7.71, 1], [14917, 748, 112, 6.68, 1], [14918, 353, 41, 8.61, 1], [14919, 1386, 194, 7.14, 1], [14920, 3384, 408, 8.29, 1], [14921, 859, 121, 7.1, 1], [14922, 1385, 175, 7.91, 1], [14923, 1228, 158, 7.77, 1], [14924, 1091, 110, 9.92, 1], [14925, 178.20000000000002, 24.200000000000003, 7.36, 0], [14926, 1253, 136, 9.21, 1], [14927, 1398.1000000000001, 166.10000000000002, 8.42, 0], [14928, 186, 27, 6.89, 1], [14929, 1070, 131, 8.17, 1], [14930, 1066, 152, 7.01, 1], [14931, 147, 22, 6.68, 1], [14932, 179, 17, 10.53, 1], [14933, 626, 58, 10.79, 1], [14934, 1319, 125, 10.55, 1], [14935, 1258, 148, 8.5, 1], [14936, 261, 41, 6.37, 1], [14937, 1399, 151, 9.26, 1], [14938, 495, 64, 7.73, 1], [14939, 948, 162, 5.85, 1], [14940, 450, 70, 6.43, 1], [14941, 1302, 181, 7.19, 1], [14942, 1536, 189, 8.13, 1], [14943, 793, 76, 10.43, 1], [14944, 2789.6000000000004, 275.0, 10.14, 0], [14945, 3308, 428, 7.73, 1], [14946, 1540.0000000000002, 181.50000000000003, 8.48, 0], [14947, 1831.5000000000002, 213.4, 8.58, 0], [14949, 4796, 484, 9.91, 1], [14950, 826.1, 89.10000000000001, 9.27, 0], [14952, 3148, 531, 5.93, 1], [14953, 1520.2, 189.20000000000002, 8.03, 0], [14954, 2025.1000000000001, 222.20000000000002, 9.11, 0], [14955, 2032.8000000000002, 193.60000000000002, 10.5, 0], [14957, 1471.8000000000002, 178.20000000000002, 8.26, 0], [14958, 1839.2, 182.60000000000002, 10.07, 0], [14959, 1124.2, 136.4, 8.24, 0], [14960, 1383.8000000000002, 167.20000000000002, 8.28, 0], [14961, 2894, 382, 7.58, 1], [14962, 1379.4, 146.3, 9.43, 0], [14963, 3451, 486, 7.1, 1], [14964, 2618, 391, 6.7, 1], [14965, 1809.5000000000002, 221.10000000000002, 8.18, 0], [14966, 2985, 419, 7.12, 1], [14967, 2393, 334, 7.16, 1], [14968, 2565, 306, 8.38, 1], [14969, 1793.0000000000002, 169.4, 10.58, 0], [14970, 3212, 405, 7.93, 1], [14971, 1079.1000000000001, 127.60000000000001, 8.46, 0], [14972, 3711, 466, 7.96, 1], [14973, 3224, 350, 9.21, 1], [14974, 2503.6000000000004, 300.3, 8.34, 0], [14975, 3031, 312, 9.71, 1], [14976, 817.3000000000001, 104.50000000000001, 7.82, 0], [14977, 1914.0000000000002, 210.10000000000002, 9.11, 0], [14978, 941.6, 132.0, 7.13, 0], [14980, 1991.0000000000002, 182.60000000000002, 10.9, 0], [14981, 1493.8000000000002, 174.9, 8.54, 0], [14982, 2072.4, 189.20000000000002, 10.95, 0], [14983, 2707.1000000000004, 223.3, 12.12, 0], [14984, 4445, 541, 8.22, 1], [14985, 1834.8000000000002, 187.00000000000003, 9.81, 0], [14986, 1824.9, 237.60000000000002, 7.68, 0], [14987, 990.0000000000001, 150.70000000000002, 6.57, 0], [14988, 1674.2, 193.60000000000002, 8.65, 0], [14989, 2646.6000000000004, 323.40000000000003, 8.18, 0], [14991, 1480.6000000000001, 155.10000000000002, 9.55, 0], [14992, 871.2, 101.2, 8.61, 0], [14993, 3726, 412, 9.04, 1], [14994, 1001.0000000000001, 134.20000000000002, 7.46, 0], [14995, 3892, 470, 8.28, 1], [14996, 1095.6000000000001, 130.9, 8.37, 0], [14997, 3541, 433, 8.18, 1], [14998, 3007.4, 301.40000000000003, 9.98, 0], [14999, 1129.7, 103.4, 10.93, 0], [15000, 855.8000000000001, 104.50000000000001, 8.19, 0], [15001, 3620, 425, 8.52, 1], [15002, 2126.3, 181.50000000000003, 11.72, 0], [15003, 3285, 355, 9.25, 1], [15004, 3251.6000000000004, 413.6, 7.86, 0], [15005, 2176.9, 222.20000000000002, 9.8, 0], [15006, 1945.9, 193.60000000000002, 10.05, 0], [15007, 831, 102, 8.15, 1], [15008, 2965.6000000000004, 371.8, 7.98, 0], [15457, 1879.9, 180.4, 10.42, 0], [15458, 2990.9, 305.8, 9.78, 0], [15459, 2094.4, 258.5, 8.1, 0], [15460, 2858, 421, 6.79, 1], [15461, 2506.9, 256.3, 9.78, 0], [15462, 1779.8000000000002, 254.10000000000002, 7.0, 0], [15463, 2200.0, 250.8, 8.77, 0], [15464, 2957, 334, 8.85, 1], [15465, 1826.0000000000002, 209.00000000000003, 8.74, 0], [15466, 2724, 347, 7.85, 1], [15467, 1597.2, 218.9, 7.3, 0], [15468, 2172, 302, 7.19, 1], [15469, 2076, 277, 7.49, 1], [15470, 1762.2, 191.4, 9.21, 0], [15471, 2997.5000000000005, 358.6, 8.36, 0], [15472, 3174, 427, 7.43, 1], [15473, 1405.8000000000002, 147.4, 9.54, 0], [15474, 2705, 308, 8.78, 1], [15475, 1896.4, 244.20000000000002, 7.77, 0], [15476, 2767, 385, 7.19, 1], [15477, 2816, 403, 6.99, 1], [15478, 2820.4, 279.40000000000003, 10.09, 0], [15479, 2025.1000000000001, 195.8, 10.34, 0], [15480, 1986, 299, 6.64, 1], [15481, 3171.3, 343.20000000000005, 9.24, 0], [15482, 1161.6000000000001, 108.9, 10.67, 0], [15483, 2316, 326, 7.1, 1], [15484, 2141.7000000000003, 199.10000000000002, 10.76, 0], [15485, 2379.3, 229.9, 10.35, 0], [15486, 2103.2000000000003, 251.90000000000003, 8.35, 0], [15487, 2347.4, 266.20000000000005, 8.82, 0], [15488, 3778.5000000000005, 375.1, 10.07, 0], [15489, 1971.2000000000003, 169.4, 11.64, 0], [15490, 2314, 293, 7.9, 1], [15491, 3219, 386, 8.34, 1], [15492, 4005, 415, 9.65, 1], [15493, 1603.8000000000002, 158.4, 10.12, 0], [15494, 1923, 271, 7.1, 1], [15495, 1824.9, 130.9, 13.94, 0], [15496, 3094, 339, 9.13, 1], [15497, 2109.8, 233.20000000000002, 9.05, 0], [15498, 3008, 345, 8.72, 1], [15499, 2339.7000000000003, 237.60000000000002, 9.85, 0], [15500, 1277.1000000000001, 149.60000000000002, 8.54, 0], [15501, 1760.0000000000002, 222.20000000000002, 7.92, 0], [15502, 3147.1000000000004, 325.6, 9.67, 0], [15503, 2161, 251, 8.61, 1], [15504, 1052.7, 134.20000000000002, 7.84, 0], [15505, 2045, 282, 7.25, 1], [15506, 1937.1000000000001, 275.0, 7.04, 0], [15507, 2099.9, 244.20000000000002, 8.6, 0], [15508, 2156, 344, 6.27, 1], [15509, 2512.4, 240.9, 10.43, 0], [15510, 1733.6000000000001, 180.4, 9.61, 0], [15511, 1965.7, 267.3, 7.35, 0], [15512, 2714, 329, 8.25, 1], [15513, 1662.1000000000001, 196.9, 8.44, 0], [15514, 2312, 314, 7.36, 1], [15515, 2431, 362, 6.72, 1], [15516, 3269, 377, 8.67, 1], [15517, 2459.6000000000004, 243.10000000000002, 10.12, 0], [15518, 2905.1000000000004, 262.90000000000003, 11.05, 0], [15519, 2718.1000000000004, 294.8, 9.22, 0], [15520, 2746, 263, 10.44, 1], [15521, 1989.9, 202.4, 9.83, 0], [15522, 2281.4, 211.20000000000002, 10.8, 0], [15523, 2082.3, 225.50000000000003, 9.23, 0], [15524, 2695, 346, 7.79, 1], [15525, 1912.9, 178.20000000000002, 10.73, 0], [15526, 2722.5, 354.20000000000005, 7.69, 0], [15527, 3371.5000000000005, 297.0, 11.35, 0], [15528, 3009.6000000000004, 308.0, 9.77, 0], [15529, 1259.5, 148.5, 8.48, 0], [15530, 2110.9, 212.3, 9.94, 0], [15531, 3723, 380, 9.8, 1], [15532, 4032.6000000000004, 403.70000000000005, 9.99, 0], [15533, 3076.7000000000003, 260.70000000000005, 11.8, 0], [15534, 1826.0000000000002, 202.4, 9.02, 0], [15535, 2548, 365, 6.98, 1], [15536, 1778.7, 163.9, 10.85, 0], [15537, 4798, 500, 9.6, 1], [15538, 2509, 345, 7.27, 1], [15539, 2795.1000000000004, 311.3, 8.98, 0], [15540, 1625.8000000000002, 191.4, 8.49, 0], [15541, 2422, 285, 8.5, 1], [15542, 1866.7, 202.4, 9.22, 0], [15543, 1915.1000000000001, 238.70000000000002, 8.02, 0], [15544, 2197, 293, 7.5, 1], [15545, 2916, 365, 7.99, 1], [15546, 1292.5, 178.20000000000002, 7.25, 0], [15547, 3034.9, 320.1, 9.48, 0], [15548, 2126, 285, 7.46, 1], [15549, 3588, 381, 9.42, 1], [15550, 3253.8, 345.40000000000003, 9.42, 0], [15551, 1620.3000000000002, 145.20000000000002, 11.16, 0], [15552, 1405.8000000000002, 169.4, 8.3, 0], [15553, 1223.2, 138.60000000000002, 8.83, 0], [15554, 2128, 309, 6.89, 1], [15555, 3092.1000000000004, 331.1, 9.34, 0], [15556, 6239, 498, 12.53, 1], [15557, 3760, 408, 9.22, 1], [15558, 2985.4, 321.20000000000005, 9.29, 0], [15559, 110.00000000000001, 1.1, 100.0, 0], [15560, 2946.9, 313.5, 9.4, 0], [15561, 2534, 299, 8.47, 1], [15562, 2702.7000000000003, 284.90000000000003, 9.49, 0], [15563, 2110.9, 243.10000000000002, 8.68, 0], [15564, 2448.6000000000004, 235.4, 10.4, 0], [15565, 2055.9, 247.50000000000003, 8.31, 0], [15566, 2928, 376, 7.79, 1], [15567, 2132.9, 227.70000000000002, 9.37, 0], [15568, 3102.0000000000005, 246.40000000000003, 12.59, 0], [15569, 3067.9, 349.8, 8.77, 0], [15570, 1844.7, 191.4, 9.64, 0], [15571, 1916.2, 244.20000000000002, 7.85, 0], [15572, 2371, 313, 7.58, 1], [15573, 3149.3, 223.3, 14.1, 0], [15574, 1035, 166, 6.23, 1], [15575, 2189.0, 259.6, 8.43, 0], [15576, 2768.7000000000003, 281.6, 9.83, 0], [15577, 1535.6000000000001, 171.60000000000002, 8.95, 0], [15578, 2990.9, 283.8, 10.54, 0], [15579, 2280.3, 260.70000000000005, 8.75, 0], [15580, 4345, 488, 8.9, 1], [15581, 1773, 207, 8.57, 1], [15582, 2026.2000000000003, 226.60000000000002, 8.94, 0], [15583, 3059, 271, 11.29, 1], [15584, 1802.9, 262.90000000000003, 6.86, 0], [15585, 2789, 333, 8.38, 1], [15586, 2500.3, 280.5, 8.91, 0], [15587, 2947, 373, 7.9, 1], [15588, 2318.8, 260.70000000000005, 8.89, 0], [15589, 2923.8, 319.0, 9.17, 0], [15590, 2248, 265, 8.48, 1], [15591, 3551.9, 389.40000000000003, 9.12, 0], [15592, 1587.3000000000002, 125.4, 12.66, 0], [15593, 1775.4, 227.70000000000002, 7.8, 0], [15594, 2379, 326, 7.3, 1], [15595, 3440, 355, 9.69, 1], [15596, 2675, 371, 7.21, 1], [15597, 3940.2000000000003, 369.6, 10.66, 0], [15598, 1656, 259, 6.39, 1], [15599, 4209, 413, 10.19, 1], [15600, 2715.9, 238.70000000000002, 11.38, 0], [15601, 2786, 325, 8.57, 1], [15602, 2072.4, 220.00000000000003, 9.42, 0], [15603, 2394.7000000000003, 250.8, 9.55, 0], [15604, 2456, 332, 7.4, 1], [15605, 3773, 423, 8.92, 1], [15606, 1344.2, 135.3, 9.93, 0], [15607, 4307.6, 412.50000000000006, 10.44, 0], [15608, 1818.3000000000002, 214.50000000000003, 8.48, 0], [15609, 1966.8000000000002, 236.50000000000003, 8.32, 0], [15610, 2478.3, 271.70000000000005, 9.12, 0], [15611, 1123.1000000000001, 134.20000000000002, 8.37, 0], [15612, 2896.3, 324.5, 8.93, 0], [15613, 2284.7000000000003, 265.1, 8.62, 0], [15614, 2675.2000000000003, 302.5, 8.84, 0], [15615, 2480, 305, 8.13, 1], [15616, 2167.0, 269.5, 8.04, 0], [15617, 2418.9, 279.40000000000003, 8.66, 0], [15618, 2361, 325, 7.26, 1], [15619, 2682, 346, 7.75, 1], [15620, 2721.4, 348.70000000000005, 7.8, 0], [15621, 3027.2000000000003, 347.6, 8.71, 0], [15622, 3939, 388, 10.15, 1], [15623, 2294, 339, 6.77, 1], [15624, 3103, 292, 10.63, 1], [15625, 3058.0000000000005, 284.90000000000003, 10.73, 0], [15626, 2542.1000000000004, 234.3, 10.85, 0], [15627, 2464.0, 234.3, 10.52, 0], [15628, 3288, 374, 8.79, 1], [15629, 2294.6000000000004, 211.20000000000002, 10.86, 0], [15630, 1805.1000000000001, 211.20000000000002, 8.55, 0], [15631, 2798, 335, 8.35, 1], [15632, 2327.6000000000004, 250.8, 9.28, 0], [15633, 1468.5000000000002, 132.0, 11.12, 0], [15634, 4499, 460, 9.78, 1], [15635, 2783.0, 310.20000000000005, 8.97, 0], [15636, 1675.3000000000002, 195.8, 8.56, 0], [15637, 3663, 364, 10.06, 1], [15638, 2627, 348, 7.55, 1], [15639, 2237, 299, 7.48, 1], [15640, 2849.0000000000005, 331.1, 8.6, 0], [15641, 1949.2, 245.3, 7.95, 0], [15642, 2501, 358, 6.99, 1], [15643, 3128.4, 378.40000000000003, 8.27, 0], [15644, 1914.0000000000002, 209.00000000000003, 9.16, 0], [15645, 2937.0000000000005, 376.20000000000005, 7.81, 0], [15646, 2717, 387, 7.02, 1], [15647, 1878.8000000000002, 228.8, 8.21, 0], [15648, 3308, 362, 9.14, 1], [15649, 2650, 366, 7.24, 1], [15650, 2996, 389, 7.7, 1], [15651, 3389, 449, 7.55, 1], [15652, 2769, 388, 7.14, 1], [15653, 2056, 252, 8.16, 1], [15654, 3814, 553, 6.9, 1], [15655, 3003, 467, 6.43, 1], [15656, 3146, 359, 8.76, 1], [15657, 2206.6000000000004, 304.70000000000005, 7.24, 0], [15658, 3462, 407, 8.51, 1], [15659, 1606.0000000000002, 213.4, 7.53, 0], [15660, 2697, 419, 6.44, 1], [15661, 3164, 470, 6.73, 1], [15662, 1860.1000000000001, 360.8, 5.16, 0], [15663, 3190, 432, 7.38, 1], [15664, 2373, 380, 6.24, 1], [15665, 2420.0, 243.10000000000002, 9.95, 0], [15666, 1548.8000000000002, 200.20000000000002, 7.74, 0], [15667, 2837, 387, 7.33, 1], [15668, 2254, 367, 6.14, 1], [15669, 866.8000000000001, 95.7, 9.06, 0], [15670, 3586, 392, 9.15, 1], [15672, 2736, 386, 7.09, 1], [15673, 2721, 411, 6.62, 1], [15674, 2102, 315, 6.67, 1], [15675, 2798, 368, 7.6, 1], [15676, 2672, 356, 7.51, 1], [15677, 2205, 338, 6.52, 1], [15678, 2995, 431, 6.95, 1], [15679, 4384, 631, 6.95, 1], [15680, 2755, 409, 6.74, 1], [15681, 3050, 423, 7.21, 1], [15682, 2747, 426, 6.45, 1], [15683, 3074, 426, 7.22, 1], [15684, 2314.4, 275.0, 8.42, 0], [15685, 3083, 453, 6.81, 1], [15686, 2065.8, 228.8, 9.03, 0], [15687, 2381, 323, 7.37, 1], [15688, 1890, 252, 7.5, 1], [15689, 4254, 564, 7.54, 1], [15690, 1366.2, 185.9, 7.35, 0], [15691, 2732, 378, 7.23, 1], [15692, 2504, 330, 7.59, 1], [15693, 3496, 570, 6.13, 1], [15694, 3117, 479, 6.51, 1], [15695, 3213, 497, 6.46, 1], [15696, 2877, 434, 6.63, 1], [15697, 2438, 354, 6.89, 1], [15698, 3196, 456, 7.01, 1], [15699, 2743, 366, 7.49, 1], [15700, 2371, 359, 6.6, 1], [15701, 2175.8, 259.6, 8.38, 0], [15702, 2644, 481, 5.5, 1], [15703, 2116.4, 240.9, 8.79, 0], [15704, 2363, 286, 8.26, 1], [15705, 3060.2000000000003, 463.1, 6.61, 0], [15706, 2765, 470, 5.88, 1], [15707, 2969, 448, 6.63, 1], [15708, 3882, 486, 7.99, 1], [15709, 2163, 418, 5.17, 1], [15710, 2417, 393, 6.15, 1], [15711, 2380.4, 253.00000000000003, 9.41, 0], [15712, 2679.6000000000004, 306.90000000000003, 8.73, 0], [15713, 2267, 356, 6.37, 1], [15714, 2017.4, 244.20000000000002, 8.26, 0], [15715, 4089.8, 448.8, 9.11, 0], [15716, 2052, 407, 5.04, 1], [15717, 2295.7000000000003, 211.20000000000002, 10.87, 0], [15718, 114.4, 15.400000000000002, 7.43, 0], [15719, 2304, 328, 7.02, 1], [15720, 3071, 490, 6.27, 1], [15721, 2003, 252, 7.95, 1], [15722, 2426, 327, 7.42, 1], [15723, 2722, 420, 6.48, 1], [15724, 2630, 417, 6.31, 1], [15725, 2359, 348, 6.78, 1], [15726, 2114, 386, 5.48, 1], [15727, 2403, 377, 6.37, 1], [15728, 2909, 391, 7.44, 1], [15729, 2832, 356, 7.96, 1], [15730, 3137, 492, 6.38, 1], [15731, 3374, 516, 6.54, 1], [15732, 2984, 354, 8.43, 1], [15733, 3121, 369, 8.46, 1], [15734, 2855, 358, 7.97, 1], [15735, 2121, 344, 6.17, 1], [15737, 1863.4, 214.50000000000003, 8.69, 0], [15738, 2143, 265, 8.09, 1], [15739, 2697.2000000000003, 302.5, 8.92, 0], [15740, 3124.0000000000005, 269.5, 11.59, 0], [15741, 3061.3, 301.40000000000003, 10.16, 0], [15742, 2024.0000000000002, 235.4, 8.6, 0], [15743, 4148, 621, 6.68, 1], [15744, 893, 121, 7.38, 1], [15745, 3761, 514, 7.32, 1], [15746, 2189, 349, 6.27, 1], [15747, 2501, 377, 6.63, 1], [15748, 2773, 411, 6.75, 1], [15749, 2395, 402, 5.96, 1], [15750, 1685, 260, 6.48, 1], [15751, 2036, 337, 6.04, 1], [15752, 3005, 452, 6.65, 1], [15753, 2784, 360, 7.73, 1], [15754, 3208, 511, 6.28, 1], [15755, 2451, 323, 7.59, 1], [15756, 3779, 489, 7.73, 1], [15757, 2949, 390, 7.56, 1], [15758, 2490, 369, 6.75, 1], [15759, 2246, 318, 7.06, 1], [15760, 2979, 380, 7.84, 1], [15761, 2061, 320, 6.44, 1], [15762, 3596, 593, 6.06, 1], [15763, 3319, 515, 6.44, 1], [15764, 3962, 635, 6.24, 1], [15765, 2641, 406, 6.5, 1], [15766, 1991.0000000000002, 238.70000000000002, 8.34, 0], [15767, 2785, 377, 7.39, 1], [15768, 3597, 511, 7.04, 1], [15769, 3249, 467, 6.96, 1], [15770, 2472.8, 363.00000000000006, 6.81, 0], [15771, 1383, 248, 5.58, 1], [15772, 3590, 424, 8.47, 1], [15773, 2807, 400, 7.02, 1], [15774, 2811, 290, 9.69, 1], [15775, 2730, 384, 7.11, 1], [15776, 3945, 440, 8.97, 1], [15777, 568.7, 126.50000000000001, 4.5, 0], [15778, 2571, 380, 6.77, 1], [15779, 2963, 435, 6.81, 1], [15780, 2324, 381, 6.1, 1], [15781, 3150, 479, 6.58, 1], [15782, 3188, 462, 6.9, 1], [15783, 2654, 356, 7.46, 1], [15784, 2042, 340, 6.01, 1], [15785, 2052, 373, 5.5, 1], [15786, 2837, 408, 6.95, 1], [15787, 3083, 422, 7.31, 1], [15788, 2640, 314, 8.41, 1], [15789, 2204, 345, 6.39, 1], [15790, 3193, 575, 5.55, 1], [15791, 3645, 575, 6.34, 1], [15792, 2209, 360, 6.14, 1], [15793, 2426, 358, 6.78, 1], [15794, 1752.3000000000002, 221.10000000000002, 7.93, 0], [15795, 3364, 425, 7.92, 1], [15796, 2149, 264, 8.14, 1], [15797, 2924, 474, 6.17, 1], [15798, 2671, 350, 7.63, 1], [15799, 2262, 377, 6.0, 1], [15800, 2093.3, 298.1, 7.02, 0], [15801, 2953.5000000000005, 345.40000000000003, 8.55, 0], [15802, 3602, 469, 7.68, 1], [15803, 2616, 365, 7.17, 1], [15804, 2671, 444, 6.02, 1], [15805, 2871, 449, 6.39, 1], [15806, 2062, 351, 5.87, 1], [15807, 3434, 350, 9.81, 1], [15808, 2589, 372, 6.96, 1], [15809, 1944, 286, 6.8, 1], [15810, 3474, 372, 9.34, 1], [15811, 2553, 426, 5.99, 1], [15812, 2632, 358, 7.35, 1], [15813, 2794, 424, 6.59, 1], [15814, 2706, 452, 5.99, 1], [15815, 3071, 441, 6.96, 1], [15816, 2850, 459, 6.21, 1], [15817, 2521, 424, 5.95, 1], [15818, 2122, 305, 6.96, 1], [15819, 3000, 383, 7.83, 1], [15820, 1871.1000000000001, 224.4, 8.34, 0], [15821, 2213, 310, 7.14, 1], [15822, 2050, 280, 7.32, 1], [15823, 2532, 442, 5.73, 1], [15824, 3134, 538, 5.83, 1], [15825, 3340, 459, 7.28, 1], [15826, 2020.7000000000003, 257.40000000000003, 7.85, 0], [15827, 2423, 385, 6.29, 1], [15828, 3465.0000000000005, 398.20000000000005, 8.7, 0], [15829, 3351.7000000000003, 279.40000000000003, 12.0, 0], [15830, 3057, 421, 7.26, 1], [15831, 4045, 520, 7.78, 1], [15832, 2609, 296, 8.81, 1], [15833, 2619, 349, 7.5, 1], [15834, 2657, 472, 5.63, 1], [15835, 2979, 451, 6.61, 1], [15836, 2542, 386, 6.59, 1], [15837, 2432, 381, 6.38, 1], [15838, 2357, 253, 9.32, 1], [15839, 3912, 436, 8.97, 1], [15840, 3350, 443, 7.56, 1], [15841, 3776.3, 392.70000000000005, 9.62, 0], [15842, 3839, 492, 7.8, 1], [15843, 2672, 428, 6.24, 1], [15844, 2099.9, 225.50000000000003, 9.31, 0], [15845, 2173.6000000000004, 258.5, 8.41, 0], [15847, 2669, 319, 8.37, 1], [15848, 1532.3000000000002, 202.4, 7.57, 0], [15849, 3408, 515, 6.62, 1], [15850, 2423.3, 319.0, 7.6, 0], [15851, 2670, 368, 7.26, 1], [15852, 3331, 510, 6.53, 1], [15853, 2566, 413, 6.21, 1], [15854, 2865, 445, 6.44, 1], [15855, 2303, 339, 6.79, 1], [15856, 3588, 521, 6.89, 1], [15857, 2339, 358, 6.53, 1], [15858, 2761, 427, 6.47, 1], [15859, 2743.4, 347.6, 7.89, 0], [15860, 2644, 456, 5.8, 1], [15861, 1031.8000000000002, 158.4, 6.51, 0], [15862, 3203, 458, 6.99, 1], [15863, 3081, 511, 6.03, 1], [15864, 2908, 429, 6.78, 1], [15865, 1648, 255, 6.46, 1], [15866, 2701, 360, 7.5, 1], [15867, 3815, 493, 7.74, 1], [15868, 2168.1000000000004, 316.8, 6.84, 0], [15869, 2203, 325, 6.78, 1], [15870, 2824, 439, 6.43, 1], [15871, 2389.2000000000003, 321.20000000000005, 7.44, 0], [15872, 3061, 370, 8.27, 1], [15873, 1684, 258, 6.53, 1], [15874, 3008, 419, 7.18, 1], [15875, 2824, 443, 6.37, 1], [15876, 2335.3, 313.5, 7.45, 0], [15877, 1908, 300, 6.36, 1], [15878, 2373, 322, 7.37, 1], [15879, 2878, 425, 6.77, 1], [15880, 1791.9, 173.8, 10.31, 0], [15881, 2232, 321, 6.95, 1], [15882, 2184, 388, 5.63, 1], [15883, 2288.0, 286.0, 8.0, 0], [15884, 2816, 391, 7.2, 1], [15885, 2692, 381, 7.07, 1], [15886, 2780, 417, 6.67, 1], [15887, 2564, 404, 6.35, 1], [15888, 3360, 466, 7.21, 1], [15889, 2917, 421, 6.93, 1], [15890, 2603, 358, 7.27, 1], [15891, 3470.5000000000005, 394.90000000000003, 8.79, 0], [15892, 2211, 330, 6.7, 1], [15893, 2700, 359, 7.52, 1], [15894, 2056, 350, 5.87, 1], [15895, 2715, 388, 7.0, 1], [15897, 4739, 531, 8.92, 1], [15898, 2875, 386, 7.45, 1], [15899, 3395, 404, 8.4, 1], [15900, 3398, 473, 7.18, 1], [15901, 3216, 439, 7.33, 1], [15902, 3249.4, 323.40000000000003, 10.05, 0], [15903, 2677, 431, 6.21, 1], [15904, 743, 86, 8.64, 1], [15905, 1821, 170, 10.71, 1], [15906, 1047, 109, 9.61, 1], [15907, 644, 117, 5.5, 1], [15908, 1823, 136, 13.4, 1], [15909, 941, 110, 8.55, 1], [15910, 448, 40, 11.2, 1], [15911, 999, 131, 7.63, 1], [15912, 690, 59, 11.69, 1], [15913, 888, 110, 8.07, 1], [15914, 541, 90, 6.01, 1], [15915, 1046, 109, 9.6, 1], [15916, 931, 96, 9.7, 1], [15917, 197, 26, 7.58, 1], [15918, 1250, 131, 9.54, 1], [15919, 1335, 152, 8.78, 1], [15920, 919, 86, 10.69, 1], [15921, 487, 65, 7.49, 1], [15922, 412, 75, 5.49, 1], [15923, 99, 9, 11.0, 1], [15924, 916, 97, 9.44, 1], [15925, 1269, 147, 8.63, 1], [15926, 680, 97, 7.01, 1], [15927, 819, 90, 9.1, 1], [15928, 942, 130, 7.25, 1], [15929, 1073, 137, 7.83, 1], [15930, 1396, 166, 8.41, 1], [15931, 867, 109, 7.95, 1], [15932, 457, 62, 7.37, 1], [15933, 882, 101, 8.73, 1], [15934, 1749, 119, 14.7, 1], [15935, 946, 127, 7.45, 1], [15936, 357, 61, 5.85, 1], [15937, 503, 52, 9.67, 1], [15938, 1611, 179, 9.0, 1], [15939, 277, 47, 5.89, 1], [15940, 1448, 138, 10.49, 1], [15941, 1691, 107, 15.8, 1], [15942, 929, 81, 11.47, 1], [15943, 499, 55, 9.07, 1], [15944, 1084, 129, 8.4, 1], [15945, 1295, 107, 12.1, 1], [15946, 1153, 137, 8.42, 1], [15947, 697, 78, 8.94, 1], [15948, 1299, 108, 12.03, 1], [15949, 941, 102, 9.23, 1], [15950, 480, 41, 11.71, 1], [15951, 821, 85, 9.66, 1], [15952, 560, 57, 9.82, 1], [15953, 785, 92, 8.53, 1], [15954, 890, 127, 7.01, 1], [15955, 979, 71, 13.79, 1], [15956, 1298, 109, 11.91, 1], [15957, 1163, 146, 7.97, 1], [15958, 1815, 150, 12.1, 1], [15959, 1735, 130, 13.35, 1], [15960, 1957, 157, 12.46, 1], [15961, 472, 50, 9.44, 1], [15962, 849, 110, 7.72, 1], [15963, 778, 87, 8.94, 1], [15964, 1508.1000000000001, 192.50000000000003, 7.83, 0], [15965, 2879.8, 418.00000000000006, 6.89, 0], [15966, 3359.4, 471.90000000000003, 7.12, 0], [15967, 3324.2000000000003, 490.6, 6.78, 0], [15968, 2784, 286, 9.73, 1], [15969, 2270.4, 372.90000000000003, 6.09, 0], [15970, 3533, 513, 6.89, 1], [15971, 1729.2, 191.4, 9.03, 0], [15972, 2903, 352, 8.25, 1], [15973, 998.8000000000001, 126.50000000000001, 7.9, 0], [15974, 3329, 499, 6.67, 1], [15975, 2673.0, 407.00000000000006, 6.57, 0], [15976, 3874, 442, 8.76, 1], [15977, 3628, 392, 9.26, 1], [15978, 2543, 352, 7.22, 1], [15979, 2741.2000000000003, 271.70000000000005, 10.09, 0], [15980, 3225, 529, 6.1, 1], [15981, 2029.5000000000002, 266.20000000000005, 7.62, 0], [15982, 2158.2000000000003, 293.70000000000005, 7.35, 0], [15983, 2911, 433, 6.72, 1], [15984, 3195, 484, 6.6, 1], [15985, 2296.8, 396.00000000000006, 5.8, 0], [15986, 2059.2000000000003, 235.4, 8.75, 0], [15987, 3020.6000000000004, 348.70000000000005, 8.66, 0], [15988, 3284.6000000000004, 521.4000000000001, 6.3, 0], [15989, 1875.5000000000002, 251.90000000000003, 7.45, 0], [15990, 3098.7000000000003, 446.6, 6.94, 0], [15991, 2635.6000000000004, 441.1, 5.98, 0], [15992, 2845.7000000000003, 372.90000000000003, 7.63, 0], [15993, 1914.0000000000002, 295.90000000000003, 6.47, 0], [15995, 2690, 390, 6.9, 1], [15996, 2533.3, 367.40000000000003, 6.9, 0], [15997, 2238, 258, 8.67, 1], [15998, 3947, 596, 6.62, 1], [15999, 2782, 276, 10.08, 1], [16000, 2823.7000000000003, 404.8, 6.98, 0], [16001, 3143, 453, 6.94, 1], [16002, 2204, 336, 6.56, 1], [16004, 1617.0000000000002, 221.10000000000002, 7.31, 0], [16005, 3360, 470, 7.15, 1], [16006, 1910.7, 209.00000000000003, 9.14, 0], [16007, 2599.3, 371.8, 6.99, 0], [16008, 3425, 411, 8.33, 1], [16009, 3387, 331, 10.23, 1], [16010, 3232, 447, 7.23, 1], [16011, 2401, 323, 7.43, 1], [16012, 3937, 360, 10.94, 1], [16014, 3022.8, 325.6, 9.28, 0], [16015, 3653, 448, 8.15, 1], [16016, 2612.5, 402.6, 6.49, 0], [16017, 1706.1000000000001, 154.0, 11.08, 0], [16018, 2362, 312, 7.57, 1], [16019, 3161, 507, 6.23, 1], [16020, 2172, 313, 6.94, 1], [16021, 2402.4, 333.3, 7.21, 0], [16022, 1953.6000000000001, 262.90000000000003, 7.43, 0], [16023, 2426, 356, 6.81, 1], [16024, 5125, 500, 10.25, 1], [16025, 3703, 477, 7.76, 1], [16026, 2960.1000000000004, 495.00000000000006, 5.98, 0], [16027, 3694, 484, 7.63, 1], [16028, 2336, 340, 6.87, 1], [16029, 3116, 398, 7.83, 1], [16030, 2676, 393, 6.81, 1], [16031, 2364, 310, 7.63, 1], [16032, 3164.7000000000003, 342.1, 9.25, 0], [16033, 2824.8, 364.1, 7.76, 0], [16034, 2682.9, 271.70000000000005, 9.87, 0], [16035, 2807, 443, 6.34, 1], [16036, 3249, 521, 6.24, 1], [16037, 3738, 505, 7.4, 1], [16038, 5080, 657, 7.73, 1], [16039, 3343, 501, 6.67, 1], [16040, 1747, 298, 5.86, 1], [16041, 2040, 255, 8.0, 1], [16042, 2958, 430, 6.88, 1], [16043, 1963, 296, 6.63, 1], [16044, 2982, 413, 7.22, 1], [16045, 2421, 367, 6.6, 1], [16046, 2456, 418, 5.88, 1], [16047, 2112, 374, 5.65, 1], [16048, 2977, 434, 6.86, 1], [16049, 2489, 385, 6.46, 1], [16050, 3097, 517, 5.99, 1], [16051, 3882, 600, 6.47, 1], [16052, 2746, 353, 7.78, 1], [16053, 2409, 299, 8.06, 1], [16054, 2814, 432, 6.51, 1], [16055, 2372, 380, 6.24, 1], [16056, 2632, 356, 7.39, 1], [16057, 2868, 457, 6.28, 1], [16058, 2448, 362, 6.76, 1], [16059, 3190.0000000000005, 398.20000000000005, 8.01, 0], [16060, 1249.6000000000001, 144.10000000000002, 8.67, 0], [16061, 2004, 246, 8.15, 1], [16062, 2341.9, 287.1, 8.16, 0], [16063, 2353, 371, 6.34, 1], [16064, 2505, 346, 7.24, 1], [16067, 2476, 393, 6.3, 1], [16068, 2413, 363, 6.65, 1], [16069, 2672, 368, 7.26, 1], [16070, 4112, 488, 8.43, 1], [16071, 1778, 246, 7.23, 1], [16072, 2076, 327, 6.35, 1], [16073, 2612, 386, 6.77, 1], [16074, 1944, 266, 7.31, 1], [16075, 3162, 460, 6.87, 1], [16076, 3397, 470, 7.23, 1], [16077, 2599, 449, 5.79, 1], [16078, 3503, 498, 7.03, 1], [16079, 2920, 379, 7.7, 1], [16080, 3085.5000000000005, 396.00000000000006, 7.79, 0], [16081, 2324.3, 253.00000000000003, 9.19, 0], [16082, 2248.4, 220.00000000000003, 10.22, 0], [16083, 1840.3000000000002, 220.00000000000003, 8.36, 0], [16084, 2304.5, 268.40000000000003, 8.59, 0], [16085, 3795, 421, 9.01, 1], [16086, 2923, 333, 8.78, 1], [16087, 1666.5000000000002, 215.60000000000002, 7.73, 0], [16088, 2713.7000000000003, 322.3, 8.42, 0], [16089, 1596.1000000000001, 147.4, 10.83, 0], [16090, 3221, 327, 9.85, 1], [16091, 4617.8, 496.1, 9.31, 0], [16092, 3074.5000000000005, 364.1, 8.44, 0], [16093, 892.1, 110.00000000000001, 8.11, 0], [16094, 3056, 357, 8.56, 1], [16095, 3000.8, 376.20000000000005, 7.98, 0], [16096, 2374.9, 298.1, 7.97, 0], [16097, 2482, 294, 8.44, 1], [16098, 1975, 262, 7.54, 1], [16099, 2358.4, 289.3, 8.15, 0], [16100, 2840.2000000000003, 360.8, 7.87, 0], [16101, 3455.1000000000004, 374.00000000000006, 9.24, 0], [16102, 2893, 385, 7.51, 1], [16103, 3033.8, 350.90000000000003, 8.65, 0], [16104, 2367.2000000000003, 256.3, 9.24, 0], [16105, 1956.9, 246.40000000000003, 7.94, 0], [16106, 3025.0000000000005, 359.70000000000005, 8.41, 0], [16107, 2058.1000000000004, 210.10000000000002, 9.8, 0], [16108, 2567.4, 218.9, 11.73, 0], [16109, 1530.1000000000001, 154.0, 9.94, 0], [16110, 4789, 337, 14.21, 1], [16111, 2086.7000000000003, 258.5, 8.07, 0], [16112, 2297, 295, 7.79, 1], [16113, 1864.5000000000002, 195.8, 9.52, 0], [16114, 3273.6000000000004, 342.1, 9.57, 0], [16115, 3189, 342, 9.32, 1], [16116, 1820.5000000000002, 220.00000000000003, 8.28, 0], [16117, 2504.7000000000003, 284.90000000000003, 8.79, 0], [16118, 4779, 426, 11.22, 1], [16119, 3053.6000000000004, 313.5, 9.74, 0], [16120, 4529.8, 474.1, 9.55, 0], [16121, 3046, 425, 7.17, 1], [16122, 3502, 439, 7.98, 1], [16123, 2913, 395, 7.37, 1], [16124, 3983, 435, 9.16, 1], [16125, 2115.3, 240.9, 8.78, 0], [16126, 2721.4, 308.0, 8.84, 0], [16127, 3910, 399, 9.8, 1], [16128, 3133.9, 353.1, 8.88, 0], [16129, 3071.2000000000003, 402.6, 7.63, 0], [16130, 2664.2000000000003, 336.6, 7.92, 0], [16131, 2823, 439, 6.43, 1], [16132, 2300.1000000000004, 233.20000000000002, 9.86, 0], [16133, 1470.7, 165.0, 8.91, 0], [16134, 2547, 317, 8.03, 1], [16135, 2965, 392, 7.56, 1], [16136, 2344, 296, 7.92, 1], [16137, 2550.9, 353.1, 7.22, 0], [16138, 3488, 439, 7.95, 1], [16139, 2266, 318, 7.13, 1], [16140, 1951, 307, 6.36, 1], [16141, 2559, 324, 7.9, 1], [16142, 2686, 304, 8.84, 1], [16143, 3206, 282, 11.37, 1], [16144, 3180, 344, 9.24, 1], [16145, 2182.4, 214.50000000000003, 10.17, 0], [16146, 4206.400000000001, 400.40000000000003, 10.51, 0], [16147, 2047.1000000000001, 243.10000000000002, 8.42, 0], [16148, 3932.5000000000005, 466.40000000000003, 8.43, 0], [16149, 1909, 305, 6.26, 1], [16150, 2945, 428, 6.88, 1], [16151, 3012, 297, 10.14, 1], [16152, 2461.8, 282.70000000000005, 8.71, 0], [16153, 1753.4, 173.8, 10.09, 0], [16154, 3337.4, 327.8, 10.18, 0], [16155, 2028.4, 248.60000000000002, 8.16, 0], [16156, 1703.9, 183.70000000000002, 9.28, 0], [16157, 2332, 383, 6.09, 1], [16158, 3517.8, 355.3, 9.9, 0], [16159, 2674.1000000000004, 400.40000000000003, 6.68, 0], [16160, 2135, 225, 9.49, 1], [16161, 1670.9, 239.8, 6.97, 0], [16162, 1929.4, 201.3, 9.58, 0]]


def discover_poverty_students():
    poverty_students = []
    for i in range(len(students_consumption)):
        if students_consumption[i][2] >= 280 and students_consumption[i][2] <= 330:
            if students_consumption[i][3] < 6.8:
                poverty_students.append(students_consumption[i])
    print("贫困学生群体：", poverty_students)
    print("贫困学生群体人数：", len(poverty_students))
    poverty_students_ID = []
    for i in range(len(poverty_students)):
        poverty_students_ID.append(poverty_students[i][0])
    print(poverty_students_ID)
    return  poverty_students_ID

# discover_poverty_students()


def discover_normal_students():
    normal_students_all = []
    normal_students = []
    for i in range(len(students_consumption)):
        if students_consumption[i][2] >280 and students_consumption[i][2] <= 330:
            if students_consumption[i][3] < 10.5 and students_consumption[i][3] > 9.5:
                normal_students_all.append(students_consumption[i])
    # print("正常学生群体：", normal_students_all)
    # print("正常学生群体人数：", len(normal_students_all))
    for i in range(20):
        normal_students.append(normal_students_all[i*2])
    print("正常学生群体：", normal_students)
    print("正常学生群体人数：", len(normal_students))
    normal_students_ID = []
    for i in range(len(normal_students)):
        normal_students_ID.append(normal_students[i][0])
    print(normal_students_ID)
    return normal_students_ID

# discover_normal_students()


def all_consumption_data():
    poverty_students_ID = discover_poverty_students()
    normal_students_ID = discover_normal_students()
    month_array = [9, 10, 11, 12, 1]
    day_array = [1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31]
    # 选取2018年9月的数据
    # consumption_data_origin = pd.read_csv(filepath_StudentsConsumption_Processed)
    # consumption_data = consumption_data_origin.drop(consumption_data_origin[consumption_data_origin['year'] != 2018].index)
    consumption_data = pd.read_csv(filepath_StudentsConsumption_Processed)
    poverty_student_consumption_month_array = []
    for i in range(len(month_array)):
        poverty_student_consumption_array = []
        consumption_data_month = consumption_data.drop(consumption_data[consumption_data['month'] != month_array[i]].index)
        for j in range(len(day_array)):
            consumption_data_day = consumption_data_month.drop(consumption_data_month[consumption_data_month['day'] != day_array[j]].index)
            for k in range(len(poverty_students_ID)):
                student_consumption_array_piece = []
                student_consumption_data = consumption_data_day.drop(consumption_data_day[consumption_data_day['bf_StudentID'] != poverty_students_ID[k]].index)
                if student_consumption_data.shape[0] > 0:
                    student_daycost_all = 0
                    for m in range(student_consumption_data.shape[0]):
                        student_daycost_all += (0 - student_consumption_data['MonDeal'].iloc[m])
                    student_consumption_array_piece.append(str(poverty_students_ID[k]))
                    student_consumption_array_piece.append(str(day_array[j]))
                    student_consumption_array_piece.append(student_daycost_all)
                else:
                    student_consumption_array_piece.append(str(poverty_students_ID[k]))
                    student_consumption_array_piece.append(str(day_array[j]))
                    student_consumption_array_piece.append(0)
                poverty_student_consumption_array.append(student_consumption_array_piece)
        poverty_student_consumption_month_array.append(poverty_student_consumption_array)
    print(poverty_student_consumption_month_array)
    print(len(poverty_student_consumption_month_array))
    print(poverty_student_consumption_month_array[0])


    normal_student_consumption_month_array = []
    for i in range(len(month_array)):
        normal_student_consumption_array = []
        consumption_data_month = consumption_data.drop(consumption_data[consumption_data['month'] != month_array[i]].index)
        for j in range(len(day_array)):
            consumption_data_day = consumption_data_month.drop(consumption_data_month[consumption_data_month['day'] != day_array[j]].index)
            for k in range(len(normal_students_ID)):
                student_consumption_array_piece = []
                student_consumption_data = consumption_data_day.drop(consumption_data_day[consumption_data_day['bf_StudentID'] != normal_students_ID[k]].index)
                if student_consumption_data.shape[0] != 0:
                    student_daycost_all = 0
                    for m in range(student_consumption_data.shape[0]):
                        student_daycost_all += (0 - student_consumption_data['MonDeal'].iloc[m])
                    student_consumption_array_piece.append(str(normal_students_ID[k]))
                    student_consumption_array_piece.append(str(day_array[j]))
                    student_consumption_array_piece.append(student_daycost_all)
                else:
                    student_consumption_array_piece.append(str(normal_students_ID[k]))
                    student_consumption_array_piece.append(str(day_array[j]))
                    student_consumption_array_piece.append(0)
                normal_student_consumption_array.append(student_consumption_array_piece)
        normal_student_consumption_month_array.append(normal_student_consumption_array)
    print(normal_student_consumption_month_array)
    print(len(normal_student_consumption_month_array))
    print(normal_student_consumption_month_array[0])
    # 保存数据
    poverty_studentID = [13948, 13992, 14135, 14306, 14410, 14416, 14501, 14551, 14568, 14570, 14611, 14695, 14734, 14824, 14836, 14862,
     15480, 15554, 15674, 15719, 15759, 15761, 15800, 15809, 15818, 15868, 15869, 15877, 15881, 15892, 15993, 16020,
     16040, 16043, 16072, 16140, 16149]

    normal_studentID = [14025, 14090, 14207, 14281, 14296, 14356, 14494, 14512, 14530, 14583, 14644, 14708, 14785, 14800, 14845, 14975, 15458, 15528, 15741, 15902]

    poverty_data = {"date": day_array, "studentID": poverty_studentID, "sept": poverty_student_consumption_month_array[0], "oct": poverty_student_consumption_month_array[1],
                    "nov": poverty_student_consumption_month_array[2], "dec": poverty_student_consumption_month_array[3], "jan": poverty_student_consumption_month_array[4]}
    normal_data = {"date": day_array, "studentID": normal_studentID, "sept": normal_student_consumption_month_array[0], "oct": normal_student_consumption_month_array[1],
                   "nov": normal_student_consumption_month_array[2], "dec": normal_student_consumption_month_array[3], "jan": normal_student_consumption_month_array[4]}
    with open('../4.Application-Level-data/PovertyHelper3.json', "w") as file:
        json.dump(poverty_data, file)
    print("完成文件加载")
    with open('../4.Application-Level-data/PovertyHelper4.json', "w") as file:
        json.dump(normal_data, file)
    print("完成文件加载")

# all_consumption_data()

# 展示学生的消费对比


def average_consumption_data():
    poverty_studentsID = [13948, 13992, 14416, 14568, 14570]
    normal_studentsID = [14025, 14090, 14207, 14281, 14296]
    consumption_data = pd.read_csv(filepath_StudentsConsumption_Processed)
    # 统计学生三个月的消费次数
    year_array = [2018]
    month_array = [9, 10, 11, 12]
    day_array = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31]
    # 产生xAxis的数据
    xAxis_data = []
    for i in range(len(month_array)):
        for j in range(len(day_array)):
            if month_array[i] == 9:
                if day_array[j] < 10:
                    date = "2018-0" + str(month_array[i]) + "-0" + str(day_array[j])
                    xAxis_data.append(date)
                else:
                    date = "2018-0" + str(month_array[i]) + "-" + str(day_array[j])
                    xAxis_data.append(date)
            else:
                if day_array[j] < 10:
                    date = "2018-" + str(month_array[i]) + "-0" + str(day_array[j])
                    xAxis_data.append(date)
                else:
                    date = "2018-" + str(month_array[i]) + "-" + str(day_array[j])
                    xAxis_data.append(date)
    print('xlabel的数据')
    print(xAxis_data)

    # 产生实际的消费次数
    poverty_students_consumeTime = []
    poverty_students_averageConsume = []
    for i in range(len(poverty_studentsID)):
        consumption_data_student = consumption_data.drop(consumption_data[consumption_data['bf_StudentID'] != poverty_studentsID[i]].index)
        poverty_student_consumeTime = []
        poverty_student_averageConsume = []
        for j in range(len(month_array)):
            student_consumption_month = consumption_data_student.drop(consumption_data_student[consumption_data_student['month'] != month_array[j]].index)
            for k in range(len(day_array)):
                student_consumption_day = student_consumption_month.drop(student_consumption_month[student_consumption_month['day'] != day_array[k]].index)
                student_consumption_piece = []
                student_average_consumption_piece = []
                poverty_index = int(j * 31 + k)
                student_consumption_piece.append(poverty_index)
                student_consumption_piece.append(student_consumption_day.shape[0])
                poverty_student_consumeTime.append(student_consumption_piece)
                student_daycost = 0
                for m in range(student_consumption_day.shape[0]):
                    student_daycost += (0 - student_consumption_day['MonDeal'].iloc[m])
                student_average_consumption_piece.append(poverty_index)
                if student_consumption_day.shape[0] > 0:
                    student_average_consumption_piece.append(int(student_daycost / student_consumption_day.shape[0]))
                else:
                    student_average_consumption_piece.append(0)
                poverty_student_averageConsume.append(student_average_consumption_piece)
        poverty_students_consumeTime.append(poverty_student_consumeTime)
        poverty_students_averageConsume.append(poverty_student_averageConsume)
    print(poverty_students_consumeTime)
    print("贫困生们次数数据量", len(poverty_students_consumeTime))
    print(poverty_students_consumeTime[0])
    print("贫困生次数数据量", len(poverty_students_consumeTime[0]))
    print(poverty_students_averageConsume)
    print("贫困生们每天平均消费金额数据量", len(poverty_students_averageConsume))
    print(poverty_students_averageConsume[0])
    print("贫困生每天平均消费金额数据量", len(poverty_students_averageConsume[0]))

    normal_students_consumeTime = []
    normal_students_averageConsume = []
    for i in range(len(normal_studentsID)):
        consumption_data_student = consumption_data.drop(consumption_data[consumption_data['bf_StudentID'] != normal_studentsID[i]].index)
        normal_student_consumeTime = []
        normal_student_averageConsume = []
        for j in range(len(month_array)):
            student_consumption_month = consumption_data_student.drop(consumption_data_student[consumption_data_student['month'] != month_array[j]].index)
            for k in range(len(day_array)):
                student_consumption_day = student_consumption_month.drop(student_consumption_month[student_consumption_month['day'] != day_array[k]].index)
                student_consumption_piece = []
                student_average_consumption_piece = []
                normal_index = int(j * 31 + k)
                student_consumption_piece.append(normal_index)
                student_consumption_piece.append(student_consumption_day.shape[0])
                normal_student_consumeTime.append(student_consumption_piece)
                student_daycost = 0
                for m in range(student_consumption_day.shape[0]):
                    student_daycost += (0 - student_consumption_day['MonDeal'].iloc[m])
                student_average_consumption_piece.append(normal_index)
                if student_consumption_day.shape[0] > 0:
                    student_average_consumption_piece.append(int(student_daycost / student_consumption_day.shape[0]))
                else:
                    student_average_consumption_piece.append(0)
                normal_student_averageConsume.append(student_average_consumption_piece)
        normal_students_consumeTime.append(normal_student_consumeTime)
        normal_students_averageConsume.append(normal_student_averageConsume)
    print(normal_students_consumeTime)
    print("非贫困生们次数数据量", len(normal_students_consumeTime))
    print(normal_students_consumeTime[0])
    print("非贫困生次数数据量", len(normal_students_consumeTime[0]))
    print(normal_students_averageConsume)
    print("非贫困生们每天平均消费金额数据量", len(normal_students_averageConsume))
    print(normal_students_averageConsume[0])
    print("非贫困生每天平均消费金额数据量", len(normal_students_averageConsume[0]))

    # 处理数据，将连续的数据转为不连续的数据，去除值为0的数据，即分段形式
    poverty_students_averageConsume_new = []
    for i in range(len(poverty_students_averageConsume)):
        poverty_students_averageConsume_new_piece = []
        for j in range(len(poverty_students_averageConsume[0])):
            if poverty_students_averageConsume[i][j][1] != 0:
                poverty_students_averageConsume_new_piece.append(poverty_students_averageConsume[i][j])
        poverty_students_averageConsume_new.append(poverty_students_averageConsume_new_piece)

    normal_students_averageConsume_new = []
    for i in range(len(normal_students_averageConsume)):
        normal_students_averageConsume_new_piece = []
        for j in range(len(normal_students_averageConsume[0])):
            if normal_students_averageConsume[i][j][1] != 0:
                normal_students_averageConsume_new_piece.append(normal_students_averageConsume[i][j])
        normal_students_averageConsume_new.append(normal_students_averageConsume_new_piece)

    # 保存数据
    poverty_data = {"date": xAxis_data, "student_1_average": poverty_students_averageConsume_new[0],
                    "student_2_average": poverty_students_averageConsume_new[1],
                    "student_3_average": poverty_students_averageConsume_new[2],
                    "student_4_average": poverty_students_averageConsume_new[3],
                    "student_5_average": poverty_students_averageConsume_new[4],
                    "student_1_time": poverty_students_consumeTime[0],
                    "student_2_time": poverty_students_consumeTime[1],
                    "student_3_time": poverty_students_consumeTime[2],
                    "student_4_time": poverty_students_consumeTime[3],
                    "student_5_time": poverty_students_consumeTime[4]}
    normal_data = {"date": xAxis_data, "student_1_average": normal_students_averageConsume_new[0],
                    "student_2_average": normal_students_averageConsume_new[1],
                    "student_3_average": normal_students_averageConsume_new[2],
                    "student_4_average": normal_students_averageConsume_new[3],
                    "student_5_average": normal_students_averageConsume_new[4],
                    "student_1_time": normal_students_consumeTime[0],
                    "student_2_time": normal_students_consumeTime[1],
                    "student_3_time": normal_students_consumeTime[2],
                    "student_4_time": normal_students_consumeTime[3],
                    "student_5_time": normal_students_consumeTime[4]}
    with open("../4.Application-Level-data/PovertyHelper5.json", "w") as file:
        json.dump(poverty_data, file)
    print("完成文件加载")
    with open("../4.Application-Level-data/PovertyHelper6.json", "w") as file:
        json.dump(normal_data, file)
    print("完成文件加载")


average_consumption_data()
